% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{aggregate}
\alias{aggregate}
\title{Aggregate}
\usage{
aggregate(x, ...)
}
\arguments{
\item{x}{Object.}

\item{...}{Additional arguments.}
}
\value{
Modified object.
}
\description{
Aggregate
}
\examples{
showMethods("aggregate")
}
