% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregateCellsToSamples.R
\name{aggregateCellsToSamples}
\alias{aggregateCellsToSamples}
\title{Aggregate cells to samples}
\usage{
aggregateCellsToSamples(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\description{
Aggregate cells to samples
}
