% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregateReplicates.R
\name{aggregateReplicates}
\alias{aggregateReplicates}
\title{Aggregate technical replicates}
\usage{
aggregateReplicates(x, ...)
}
\arguments{
\item{x}{Object.}

\item{...}{Additional arguments.}
}
\value{
Modified object of same class, with columns aggregated.
}
\description{
Aggregate technical replicates
}
\examples{
class(aggregateReplicates)
}
