% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha.R
\name{alpha}
\alias{alpha}
\alias{alpha<-}
\title{Alpha level}
\usage{
alpha(object, ...)

alpha(object, ...) <- value
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{value}{Value.}
}
\description{
Alpha level
}
