% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alphaSummary.R
\name{alphaSummary}
\alias{alphaSummary}
\title{Alpha level cutoff summary statistics}
\usage{
alphaSummary(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\verb{integer matrix}.
}
\description{
Quickly generate a summary table of various alpha level cutoffs.
}
\examples{
showMethods("alphaSummary")
}
