% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha.R
\name{alphaThreshold}
\alias{alphaThreshold}
\alias{alphaThreshold<-}
\title{Alpha threshold}
\usage{
alphaThreshold(object, ...)

alphaThreshold(object, ...) <- value
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{value}{Value.}
}
\description{
Alpha threshold
}
