% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioconductor-02-generics.R
\name{as.data.frame}
\alias{as.data.frame}
\title{Coerce object to a data frame}
\usage{
as.data.frame(x, row.names = NULL, optional = FALSE, ...)
}
\arguments{
\item{x, row.names, optional, ...}{See \code{\link[base:as.data.frame]{as.data.frame()}} for details.}
}
\value{
\code{data.frame}.
}
\description{
Coerce object to a data frame
}
\examples{
showMethods("as.data.frame")
}
