% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.DESeqDataSet.R
\name{as.DESeqDataSet}
\alias{as.DESeqDataSet}
\title{Coerce object to \code{DESeqDataSet}}
\usage{
as.DESeqDataSet(x, ...)
}
\arguments{
\item{x}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{DESeqDataSet}.
}
\description{
Coerce object to \code{DESeqDataSet}
}
\examples{
showMethods("as.DESeqDataSet")
}
