% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.DESeqTransform.R
\name{as.DESeqTransform}
\alias{as.DESeqTransform}
\title{Coerce object to \code{DESeqTransform}}
\usage{
as.DESeqTransform(x, ...)
}
\arguments{
\item{x}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{DESeqTransform}.
}
\description{
Coerce object to \code{DESeqTransform}
}
\examples{
showMethods("as.DESeqTransform")
}
