% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.DFrame.R
\name{as.DFrame}
\alias{as.DFrame}
\title{Coerce object to S4 DFrame (DataFrame)}
\usage{
as.DFrame(x, ...)
}
\arguments{
\item{x}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{DFrame}.
}
\description{
Coerce object to S4 DFrame (DataFrame)
}
\examples{
showMethods("as.DFrame")
}
