% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.DGEList.R
\name{as.DGEList}
\alias{as.DGEList}
\title{Coerce object to \code{DGEList}}
\usage{
as.DGEList(x, ...)
}
\arguments{
\item{x}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{DGEList}.
}
\description{
Coerce object to \code{DGEList}
}
\examples{
showMethods("as.DGEList")
}
