% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.list.R
\name{as.list}
\alias{as.list}
\title{Coerce object to a list}
\usage{
as.list(x, ...)
}
\arguments{
\item{x}{Object.}

\item{...}{Additional arguments.}
}
\description{
Coerce object to a list
}
\note{
Updated 2020-12-12.
}
