% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.Seurat.R
\name{as.Seurat}
\alias{as.Seurat}
\title{Coerce object to Seurat}
\usage{
as.Seurat(x, ...)
}
\arguments{
\item{x}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{Seurat}.
}
\description{
Coerce object to Seurat
}
