% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.SingleCellExperiment.R
\name{as.SingleCellExperiment}
\alias{as.SingleCellExperiment}
\title{Coerce object to \code{SingleCellExperiment}}
\usage{
as.SingleCellExperiment(x, ...)
}
\arguments{
\item{x}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{DESeqDataSet}.
}
\description{
Coerce object to \code{SingleCellExperiment}
}
\examples{
showMethods("as.SingleCellExperiment")
}
