% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.SummarizedExperiment.R
\name{as.SummarizedExperiment}
\alias{as.SummarizedExperiment}
\title{Coerce object to SummarizedExperiment}
\usage{
as.SummarizedExperiment(x, ...)
}
\arguments{
\item{x}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{SummarizedExperiment}.
}
\description{
Coerce object to SummarizedExperiment
}
\examples{
class(as.SummarizedExperiment)
}
