% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atomize.R
\name{atomize}
\alias{atomize}
\title{Atomize}
\usage{
atomize(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
Modified object.
Non-atomic columns (e.g. \code{list} or complex S4 classes) will be dropped.
}
\description{
Ensure all columns in the object are \code{atomic}.
}
