% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autopadZeros.R
\name{autopadZeros}
\alias{autopadZeros}
\title{Autopad zeros}
\usage{
autopadZeros(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
Modified object, with padded zeros.
}
\description{
Autopad zeros
}
\examples{
showMethods("autopadZeros")
}
