% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barcodeRanksPerSample.R
\name{barcodeRanksPerSample}
\alias{barcodeRanksPerSample}
\title{Barcode ranks per sample}
\usage{
barcodeRanksPerSample(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{list}.
}
\description{
Barcode ranks per sample
}
\examples{
class(barcodeRanksPerSample)
}
