% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseMeanThreshold.R
\name{baseMeanThreshold}
\alias{baseMeanThreshold}
\alias{baseMeanThreshold<-}
\title{Base mean threshold}
\usage{
baseMeanThreshold(object, ...)

baseMeanThreshold(object, ...) <- value
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{value}{Value to assign.}
}
\value{
\code{numeric(1)}.
Base mean cutoff level.
}
\description{
Base mean threshold
}
\examples{
class(baseMeanThreshold)
}
