% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateMetrics.R
\name{calculateMetrics}
\alias{calculateMetrics}
\title{Calculate quality control metrics}
\usage{
calculateMetrics(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{DataFrame}.
}
\description{
Calculate quality control metrics
}
\examples{
showMethods("calculateMetrics")
}
