% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cast.R
\name{cast}
\alias{cast}
\title{Cast data from long to wide format}
\usage{
cast(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{DataFrame}.
}
\description{
Pivot column data from long format to wide format.
}
\examples{
showMethods("cast")
}
\seealso{
\itemize{
\item https://seananderson.ca/2013/10/19/reshape/
\item \code{reshape2::dcast} (deprecated).
\item \code{tidyr::pivot_wider}.
\item Python \code{pandas.DataFrame.pivot}.
\item Python \code{pandas.pivot_table}.
}
}
