% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell2sample.R
\name{cell2sample}
\alias{cell2sample}
\title{Cell-to-sample mappings}
\usage{
cell2sample(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{factor}.
}
\description{
Cell-to-sample mappings
}
\examples{
showMethods("cell2sample")
}
