% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellsPerGeneFusion.R
\name{cellsPerGeneFusion}
\alias{cellsPerGeneFusion}
\title{Cells per gene fusion}
\usage{
cellsPerGeneFusion(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{DFrame} or \code{NULL}.
Gene fusion pairs in columns, cells in rows.
Includes additional cell line metadata on left side.
Returns \code{NULL} if no cells match or pass \code{minCells} filter.
}
\description{
Cells per gene fusion
}
\examples{
showMethods("cellsPerGeneFusion")
}
