% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellsPerMutation.R
\name{cellsPerMutation}
\alias{cellsPerMutation}
\title{Cells per mutation}
\usage{
cellsPerMutation(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{DFrame} or \code{NULL}.
Genes in columns, cells in rows.
Includes additional cell line metadata on left side.
Returns \code{NULL} if no cells match or pass \code{minCells} filter.
}
\description{
Cells per mutation
}
\examples{
showMethods("cellsPerMutation")
}
