% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellToSample.R
\name{cellToSample}
\alias{cellToSample}
\title{Cell-to-sample mappings}
\usage{
cellToSample(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{factor}.
}
\description{
Cell-to-sample mappings
}
\examples{
showMethods("cellToSample")
}
