% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellTypesPerCluster.R
\name{cellTypesPerCluster}
\alias{cellTypesPerCluster}
\title{Cell types per cluster}
\usage{
cellTypesPerCluster(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{DataFrame}.
}
\description{
Cell types per cluster
}
\examples{
showMethods("cellTypesPerCluster")
}
