% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterID.R
\name{clusterID}
\alias{clusterID}
\title{Cluster identifier}
\usage{
clusterID(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{factor}.
}
\description{
Cluster identifier
}
\examples{
class(clusterID)
}
