% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusters.R
\name{clusters}
\alias{clusters}
\title{Cluster identifiers}
\usage{
clusters(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{factor}.
}
\description{
Cluster identifiers
}
\examples{
class(clusters)
}
