% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioconductor-02-generics.R
\name{coerce}
\alias{coerce}
\title{Coerce an object to a class}
\usage{
coerce(from, to, strict = TRUE)
}
\arguments{
\item{from, to, strict}{See \code{\link[methods:setAs]{coerce}} for details.}
}
\value{
Modified object of new desired class.
}
\description{
Coerce an object to a class
}
\examples{
showMethods("coerce")
}
