% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coerceToList.R
\name{coerceToList}
\alias{coerceToList}
\title{Coerce to list}
\usage{
coerceToList(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{list}.
}
\description{
Coerce to list
}
\examples{
showMethods("coerceToList")
}
