% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectionNames.R
\name{collectionNames}
\alias{collectionNames}
\alias{collectionNames<-}
\title{Gene set collection names}
\usage{
collectionNames(object, ...)

collectionNames(object, ...) <- value
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{value}{Value to assign.}
}
\value{
\code{character}.
}
\description{
Gene set collection names
}
\examples{
showMethods("collectionNames")
}
