% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colSums.R
\name{colSums}
\alias{colSums}
\title{Column sums}
\usage{
colSums(x, na.rm = FALSE, dims = 1, ...)
}
\arguments{
\item{x, na.rm, dims, ...}{See \code{\link[base:colSums]{base::colSums}} for details.}
}
\description{
Column sums
}
\note{
Updated 2020-12-21.
}
