% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.R
\name{combine}
\alias{combine}
\title{Combine multiple objects}
\usage{
combine(x, y, ...)
}
\arguments{
\item{x}{Primary object.}

\item{y}{Secondary object.}

\item{...}{Additional arguments.}
}
\description{
Combine multiple objects
}
\note{
Updated 2020-12-21.
}
