% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complete.cases.R
\name{complete.cases}
\alias{complete.cases}
\title{Complete cases}
\usage{
complete.cases(...)
}
\arguments{
\item{...}{A sequence of vectors, matrices, or data frames.}
}
\value{
\code{logical}.
}
\description{
Complete cases
}
\examples{
showMethods("complete.cases")
}
