% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrastName.R
\name{contrastName}
\alias{contrastName}
\alias{contrastName<-}
\title{Contrast name}
\usage{
contrastName(object, ...)

contrastName(object, ...) <- value
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{value}{Value to assign.}
}
\value{
\code{character(1)}.
Contrast name.
}
\description{
Contrast name
}
\examples{
class(contrastName)
}
