% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrastNames.R
\name{contrastNames}
\alias{contrastNames}
\alias{contrastNames<-}
\title{Contrast names}
\usage{
contrastNames(object, ...)

contrastNames(object, ...) <- value
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{value}{Value.}
}
\description{
Contrast names
}
