% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrastSamples.R
\name{contrastSamples}
\alias{contrastSamples}
\title{Samples corresponding to a differential expression contrast}
\usage{
contrastSamples(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\description{
Samples corresponding to a differential expression contrast
}
