% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertSampleIDsToNames.R
\name{convertSampleIDsToNames}
\alias{convertSampleIDsToNames}
\title{Convert sample identifiers to names}
\usage{
convertSampleIDsToNames(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\description{
Convert sample identifiers to names
}
