% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertSymbolsToGenes.R
\name{convertSymbolsToGenes}
\alias{convertSymbolsToGenes}
\title{Convert symbols to genes}
\usage{
convertSymbolsToGenes(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\description{
Convert symbols to genes
}
