% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertToHuman.R
\name{convertToHuman}
\alias{convertToHuman}
\title{Convert data set by mapping to human orthologs}
\usage{
convertToHuman(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
Modified object.
Features (i.e. rownames) will be remapped to human genes.
}
\description{
Convert data set by mapping to human orthologs
}
\examples{
class(convertToHuman)
}
