% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertTranscriptsToGenes.R
\name{convertTranscriptsToGenes}
\alias{convertTranscriptsToGenes}
\title{Convert transcripts to genes}
\usage{
convertTranscriptsToGenes(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
Modified object of same class, with transcripts remapped to genes.
}
\description{
Convert transcripts to genes
}
\examples{
class(convertTranscriptsToGenes)
}
