% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor.R
\name{cor}
\alias{cor}
\title{Correlation}
\usage{
cor(
    x,
    y = NULL,
    use = "everything",
    method = c("pearson", "kendall", "spearman")
)
}
\arguments{
\item{x, y, use, method}{See \code{\link[stats:cor]{stats::cor}} for details.}
}
\value{
\code{numeric(1)} or \code{matrix}.
}
\description{
Correlation
}
\examples{
showMethods("cor")
}
