% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation.R
\name{correlation}
\alias{correlation}
\title{Correlation}
\usage{
correlation(x, y, ...)
}
\arguments{
\item{x}{Object.}

\item{y}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{numeric(1)} or \code{matrix}.
}
\description{
Correlation
}
\examples{
showMethods("correlation")
}
\seealso{
\itemize{
\item \code{stats::cor()}.
\item \code{stats::cor.test()}.
\item \code{stats::sd()}.
\item https://stats.stackexchange.com/questions/24980
}
}
