% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpm.R
\name{cpm}
\alias{cpm}
\title{Counts per million}
\usage{
cpm(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{matrix}
}
\description{
Counts per million
}
\examples{
showMethods("cpm")
}
