% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{defunct}
\alias{defunct}
\alias{plotGene}
\title{Defunct functions}
\usage{
plotGene(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{.Defunct}.
}
\description{
Defunct functions
}
\seealso{
\link[base:Defunct]{Defunct}.
}
\keyword{internal}
