% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/degIntersection.R
\name{degIntersection}
\alias{degIntersection}
\title{Intersection of differentially expressed genes}
\usage{
degIntersection(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{integer}.
Integer denoting the number of intersections, sorted from highest to
lowest. Gene identifiers are defined as the names.
}
\description{
Intersection of differentially expressed genes
}
\examples{
showMethods("degIntersection")
}
