% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/degPerContrast.R
\name{degPerContrast}
\alias{degPerContrast}
\title{Differentially expressed genes per contrast}
\usage{
degPerContrast(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{list}.
Gene identifiers per contrast.
}
\description{
Differentially expressed genes per contrast
}
\examples{
showMethods("degPerContrast")
}
