% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffExp.R
\name{diffExp}
\alias{diffExp}
\title{Differential expression}
\usage{
diffExp(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
Varies, depending on the method.
}
\description{
Differential expression
}
\examples{
class(diffExp)
}
