% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do.call.R
\name{do.call}
\alias{do.call}
\title{Execute a function call}
\usage{
do.call(what, args, quote = FALSE, envir = parent.frame())
}
\arguments{
\item{what, args, quote, envir}{See \code{\link[base:do.call]{base::do.call}} for details.}
}
\value{
Varies, depending on function called.
}
\description{
Execute a function call
}
\examples{
showMethods("do.call")
}
