% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dotted.R
\name{dotted}
\alias{dotted}
\title{Dotted case}
\usage{
dotted(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
Modified object, with names formatted in dotted case.
}
\description{
Dotted case
}
\examples{
class(dotted)
}
