% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dottedCase.R
\name{dottedCase}
\alias{dottedCase}
\title{Dotted case}
\usage{
dottedCase(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
Modified object.
Contains syntatically valid names.
For objects with \code{\link[base:names]{names()}} defined, the underlying data
returns unchanged, except for \code{character} or \code{vector} class.
}
\description{
Support is provided for matching against base R parameters. However, it is
recommended to avoid using it for variable assignments into an \code{environment},
as that can introduce conflicts with base functions.
}
\examples{
class(dottedCase)
}
