% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/droplevels2.R
\name{droplevels2}
\alias{droplevels2}
\title{Drop factor levels}
\usage{
droplevels2(x, ...)
}
\arguments{
\item{x}{Object.}

\item{...}{Additional arguments.}
}
\value{
Modified object.
}
\description{
Drop factor levels
}
\examples{
showMethods("droplevels2")
}
