% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encode.R
\name{encode}
\alias{encode}
\title{Encode using run-length encoding}
\usage{
encode(x, ...)
}
\arguments{
\item{x}{Object.}

\item{...}{Additional arguments.}
}
\value{
Modified object, with data encoded into \code{Rle} class.
}
\description{
Encode using run-length encoding
}
\examples{
class(encode)
}
