% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichedGeneSets.R
\name{enrichedGeneSets}
\alias{enrichedGeneSets}
\title{Enriched gene sets}
\usage{
enrichedGeneSets(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\description{
Enriched gene sets
}
