% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ensembl2Entrez.R
\name{Ensembl2Entrez}
\alias{Ensembl2Entrez}
\title{Ensembl-to-Entrez gene identifier mappings}
\usage{
Ensembl2Entrez(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{Ensembl2Entrez}.
}
\description{
Ensembl-to-Entrez gene identifier mappings
}
\examples{
showMethods("Ensembl2Entrez")
}
