% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ensembl2Ncbi.R
\name{Ensembl2Ncbi}
\alias{Ensembl2Ncbi}
\alias{Ensembl2Entrez}
\title{Ensembl-to-NCBI gene identifier mappings}
\usage{
Ensembl2Ncbi(object, ...)

Ensembl2Entrez(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{Ensembl2Ncbi}.
}
\description{
Ensembl-to-NCBI gene identifier mappings
}
\examples{
showMethods("Ensembl2Ncbi")
}
