% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnsemblToNcbi.R
\name{EnsemblToNcbi}
\alias{EnsemblToNcbi}
\title{Ensembl-to-NCBI gene identifier mappings}
\usage{
EnsemblToNcbi(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{EnsemblToNcbi}.
}
\description{
Ensembl-to-NCBI gene identifier mappings
}
\examples{
showMethods("EnsemblToNcbi")
}
