% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Entrez2Ensembl.R
\name{Entrez2Ensembl}
\alias{Entrez2Ensembl}
\title{Entrez-to-Ensembl gene identifier mappings}
\usage{
Entrez2Ensembl(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{Entrez2Ensembl}.
}
\description{
Entrez-to-Ensembl gene identifier mappings
}
\examples{
showMethods("Entrez2Ensembl")
}
