% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateSizeFactors.R
\name{estimateSizeFactors}
\alias{estimateSizeFactors}
\title{Estimate size factors}
\usage{
estimateSizeFactors(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\description{
Estimate size factors
}
\note{
Updated 2020-12-21.
}
