% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidean.R
\name{euclidean}
\alias{euclidean}
\title{Calculate Euclidean distance}
\usage{
euclidean(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
Numeric vector.
}
\description{
Calculate Euclidean distance
}
\examples{
showMethods("euclidean")
}
