% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excludeContaminatedCells.R
\name{excludeContaminatedCells}
\alias{excludeContaminatedCells}
\title{Exclude contaminated cells}
\usage{
excludeContaminatedCells(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
Modified object, with contaminated cells excluded (removed).
}
\description{
Exclude contaminated cells
}
\examples{
showMethods("excludeContaminatedCells")
}
\seealso{
\itemize{
\item Cellosaurus package.
}
}
