% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excludeNonCancerCells.R
\name{excludeNonCancerCells}
\alias{excludeNonCancerCells}
\title{Exclude non-cancer cell lines}
\usage{
excludeNonCancerCells(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
Modified object, with non-cancer cells excluded (removed).
}
\description{
Exclude non-cancer cell lines
}
\examples{
showMethods("excludeNonCancerCells")
}
