% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excludeNonHumanCells.R
\name{excludeNonHumanCells}
\alias{excludeNonHumanCells}
\title{Exclude non-human cell lines}
\usage{
excludeNonHumanCells(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
Modified object, with non-human cells excluded (removed).
}
\description{
Exclude non-human cell lines
}
\examples{
showMethods("excludeNonHumanCells")
}
