% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excludeProblematicCells.R
\name{excludeProblematicCells}
\alias{excludeProblematicCells}
\title{Exclude problematic cells}
\usage{
excludeProblematicCells(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
Modified object, with problematic cells excluded (removed).
}
\description{
Exclude problematic cells
}
\examples{
showMethods("excludeProblematicCells")
}
\seealso{
\itemize{
\item Cellosaurus package.
}
}
