% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand.grid.R
\name{expand.grid}
\alias{expand.grid}
\title{Create a data frame from all combinations of factor variables}
\usage{
expand.grid(..., KEEP.OUT.ATTRS = TRUE, stringsAsFactors = TRUE)
}
\arguments{
\item{..., KEEP.OUT.ATTRS, stringsAsFactors}{See \code{\link[base:expand.grid]{expand.grid}} for details.}
}
\value{
\code{data.frame}.
}
\description{
Create a data frame from all combinations of factor variables
}
\examples{
showMethods("expand.grid")
}
