% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{export}
\alias{export}
\title{Export data from R}
\usage{
export(object, con, ...)
}
\arguments{
\item{object}{Object.}

\item{con}{\code{character(1)} or \code{connection}.
Data connection. Use \code{character(1)} to represent a file path.}

\item{...}{Additional arguments.}
}
\value{
Invisible \code{character} or \code{list} of file paths.
}
\description{
Export data from R
}
\examples{
showMethods("export")
}
\seealso{
\itemize{
\item \code{BiocIO::export}.
\item \code{rio::export}.
\item \code{write.table}.
\item \code{writeLines}.
}
}
