% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorize.R
\name{factorize}
\alias{factorize}
\title{Factorize}
\usage{
factorize(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
Modified object.
All columns will be coerced to \code{factor}.
}
\description{
Coerce all columns in the object to \code{factor} (i.e. \code{stringsAsFactors}).
}
\seealso{
\code{\link[=encode]{encode()}} for run-length encoding (Rle) approach.
}
