% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterNested.R
\name{filterNested}
\alias{filterNested}
\title{Keep rows that match a nested condition}
\usage{
filterNested(object, pattern, ...)
}
\arguments{
\item{object}{Object.}

\item{pattern}{\code{character(1)}.
Pattern to use for matching.}

\item{...}{Additional arguments.}
}
\value{
Modified object, containing only matching rows.
}
\description{
Keep rows that match a nested condition
}
\details{
Intended for partial matching with regular expressions. For exact matches,
refer to \code{matchNested} function.
}
\examples{
showMethods("filterNested")
}
