% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findMarkers.R
\name{findMarkers}
\alias{findMarkers}
\title{Find cluster-specific marker genes}
\usage{
findMarkers(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{data.frame}.
}
\description{
Find cluster-specific marker genes
}
\examples{
showMethods("findMarkers")
}
