% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flatFiles.R
\name{flatFiles}
\alias{flatFiles}
\title{Coerce any S4 object to a list}
\usage{
flatFiles(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\description{
Coerce any S4 object to a list
}
