% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foldChangeToLogRatio.R
\name{foldChangeToLogRatio}
\alias{foldChangeToLogRatio}
\title{Fold change to log ratio}
\usage{
foldChangeToLogRatio(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\description{
Fold change to log ratio
}
