% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpkm.R
\name{fpkm}
\alias{fpkm}
\title{Fragments per kilobase per million mapped fragments}
\usage{
fpkm(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{matrix}.
}
\description{
Return fragment counts normalized per kilobase of feature length per million
mapped fragments.
}
\note{
Only recommended for gene-level counts.
}
\seealso{
\itemize{
\item \code{DESeq2::fpkm()}.
}
}
