% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gather.R
\name{gather}
\alias{gather}
\title{Gather columns into key-value pairs}
\usage{
gather(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{DataFrame}.
}
\description{
Gather columns into key-value pairs
}
\examples{
class(gather)
}
\seealso{
\code{tidyr::gather()}.
}
