% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gene2Symbol.R
\name{Gene2Symbol}
\alias{Gene2Symbol}
\title{Gene-to-symbol mappings}
\usage{
Gene2Symbol(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{Gene2Symbol}.
}
\description{
Gene-to-symbol mappings
}
\examples{
showMethods("Gene2Symbol")
}
