% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneFusions.R
\name{geneFusions}
\alias{geneFusions}
\title{Gene fusion pairs}
\usage{
geneFusions(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{CharacterList}.
}
\description{
Gene fusion pairs
}
\examples{
showMethods("geneFusions")
}
