% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneIds.R
\name{geneIds}
\alias{geneIds}
\title{Gene identifiers}
\usage{
geneIds(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{character}.
}
\description{
Convenience function for quickly returning the stable gene identifiers.
}
\examples{
showMethods("geneIds")
}
