% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneNames.R
\name{geneNames}
\alias{geneNames}
\title{Gene names}
\usage{
geneNames(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{character}.
}
\description{
Convenience function for quickly returning gene names
(a.k.a gene symbols) that map to the stable gene identifiers.
}
\examples{
showMethods("geneNames")
}
