% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneSet.R
\name{geneSet}
\alias{geneSet}
\title{Gene set}
\usage{
geneSet(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{character}.
}
\description{
Gene set
}
\examples{
showMethods("geneSet")
}
