% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneSetNames.R
\name{geneSetNames}
\alias{geneSetNames}
\title{Gene set names}
\usage{
geneSetNames(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{character}.
}
\description{
Gene set names
}
\examples{
showMethods("geneSetNames")
}
