% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneSetResults.R
\name{geneSetResults}
\alias{geneSetResults}
\title{Gene set results}
\usage{
geneSetResults(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{DataFrame}.
}
\description{
Gene set results
}
\examples{
class(geneSetResults)
}
