% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneToSymbol.R
\name{GeneToSymbol}
\alias{GeneToSymbol}
\title{Gene-to-symbol mappings}
\usage{
GeneToSymbol(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{GeneToSymbol}.
}
\description{
Gene-to-symbol mappings
}
\examples{
showMethods("GeneToSymbol")
}
