% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometricMean.R
\name{geometricMean}
\alias{geometricMean}
\title{Geometric mean}
\usage{
geometricMean(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\description{
Geometric mean
}
