% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioconductor-02-generics.R
\name{head}
\alias{head}
\title{Return the first part of an object}
\usage{
head(x, ...)
}
\arguments{
\item{x}{Object.}

\item{...}{Additional arguments.}
}
\value{
Console output.
}
\description{
Return the first part of an object
}
\examples{
showMethods("head")
}
