% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/headtail.R
\name{headtail}
\alias{headtail}
\title{Return the first and last parts of an object}
\usage{
headtail(x, ...)
}
\arguments{
\item{x}{Object.}

\item{...}{Additional arguments.}
}
\value{
Invisible \code{NULL}.
}
\description{
Inspired by the \code{print} method for \code{DataFrame} class objects. Applies to both
rows and columns, enabling quick inspection during interactive use.
}
\seealso{
\itemize{
\item \code{utils::head()}, \code{utils::tail()}, \code{cat()}.
\item \code{getMethod("show", "DataTable")}.
}
}
