% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/humanize.R
\name{humanize}
\alias{humanize}
\title{Humanize an R object}
\usage{
humanize(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\description{
Humanize an R object
}
\note{
This can make dimnames invalid (see \code{make.names}) and should only be
called prior to writing files to disk.
}
