% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import}
\alias{import}
\title{Import data into R}
\usage{
import(con, ...)
}
\arguments{
\item{con}{\code{character(1)} (recommended) or \code{connection} (edge case).
Data connection. Most commonly, use \code{character(1)} to represent a file path
or URL. Less commonly, can create a \code{textConnection} to a character vector
of source code lines (text), which is useful for reformatting malformed files
directly in R.}

\item{...}{Additional arguments.}
}
\value{
Object.
}
\description{
Import data into R
}
\examples{
showMethods("import")
}
\seealso{
\itemize{
\item \code{BiocIO::import}.
\item \code{rio::import}.
\item \code{read.table}.
\item \code{readLines}.
\item \code{textConnection}.
}
}
