% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integerCounts.R
\name{integerCounts}
\alias{integerCounts}
\title{Integer counts}
\usage{
integerCounts(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
Matrix.
Typically \code{matrix} or \code{Matrix} class.
}
\description{
Integer counts
}
\examples{
showMethods("integerCounts")
}
