% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interestingGroups.R
\name{interestingGroups}
\alias{interestingGroups}
\alias{interestingGroups<-}
\title{Interesting groups}
\usage{
interestingGroups(object, ...)

interestingGroups(object, ...) <- value
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{value}{Value.}
}
\value{
\code{character}.
}
\description{
Interesting groups
}
\examples{
class(interestingGroups)
}
