% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intersectAll.R
\name{intersectAll}
\alias{intersectAll}
\title{Intersection of more than two vectors}
\usage{
intersectAll(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{vector}.
}
\description{
Intersection of more than two vectors
}
\examples{
class(intersectAll)
}
