% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intersectionMatrix.R
\name{intersectionMatrix}
\alias{intersectionMatrix}
\title{Intersection matrix}
\usage{
intersectionMatrix(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{matrix}.
}
\description{
Intersection matrix
}
\examples{
showMethods("intersectionMatrix")
}
