% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join.R
\name{join}
\alias{join}
\alias{innerJoin}
\alias{leftJoin}
\alias{rightJoin}
\alias{fullJoin}
\alias{semiJoin}
\alias{antiJoin}
\title{Join operations}
\usage{
innerJoin(x, y, by, ...)

leftJoin(x, y, by, ...)

rightJoin(x, y, by, ...)

fullJoin(x, y, by, ...)

semiJoin(x, y, by, ...)

antiJoin(x, y, by, ...)
}
\arguments{
\item{x}{Object.}

\item{y}{Object.}

\item{by}{\code{character}.
Column names to use for join operation.}

\item{...}{Additional arguments.}
}
\value{
New object containing a merge of \code{x} and \code{y} objects.
}
\description{
Join operations
}
\examples{
class(leftJoin)
}
\seealso{
\itemize{
\item \code{help(topic = "join", package = "dplyr")}.
\item \code{help(topic = "merge", package = "S4Vectors")}.
\item https://support.bioconductor.org/p/120277/
}
}
