% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kebab.R
\name{kebab}
\alias{kebab}
\title{Kebab case}
\usage{
kebab(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\description{
Format character strings to use kebab-style formatting, where word boundaries
are defined by dashes/hyphens (e.g. \code{this-is-kebab-case}).
}
\details{
Kebab case is recommended for URL slugs and file names. However, they should
not be used for names in R, since dashes are not valid, and should be
substituted with underscores or periods instead.
}
\seealso{
\code{\link[base:make.names]{make.names()}}.
}
