% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keepOnlyAtomicCols.R
\name{keepOnlyAtomicCols}
\alias{keepOnlyAtomicCols}
\title{Keep only atomic columns}
\usage{
keepOnlyAtomicCols(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
Modified object, containing only atomic columns.
}
\description{
Keep only atomic columns
}
\examples{
showMethods("keepOnlyAtomicCols")
}
