% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leadingEdge.R
\name{leadingEdge}
\alias{leadingEdge}
\title{Leading edge genes}
\usage{
leadingEdge(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{character}.
Elements of gene set that correspond to leading edge.
}
\description{
Leading edge genes
}
\examples{
class(leadingEdge)
}
