% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfcShrink.R
\name{lfcShrink}
\alias{lfcShrink}
\alias{lfcShrink<-}
\title{Shrink log2 fold changes}
\usage{
lfcShrink(object, ...)

lfcShrink(object, ...) <- value
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{value}{Value to assign.}
}
\value{
Varies, depending on object class.
}
\description{
Shrink log2 fold changes
}
\examples{
class(lfcShrink)
}
