% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfcShrinkType.R
\name{lfcShrinkType}
\alias{lfcShrinkType}
\title{Shrunken log2 fold change (LFC) type}
\usage{
lfcShrinkType(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{character(1)} or \code{NULL}.
Name of adaptive shrinkage method applied (e.g. "apeglm" or "ashr").
}
\description{
Shrunken log2 fold change (LFC) type
}
\examples{
showMethods("lfcShrinkType")
}
