% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfcThreshold.R
\name{lfcThreshold}
\alias{lfcThreshold}
\alias{lfcThreshold<-}
\title{Log2 fold change threshold}
\usage{
lfcThreshold(object, ...)

lfcThreshold(object, ...) <- value
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{value}{Value.}
}
\description{
Log2 fold change threshold
}
