% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logRatioToFoldChange.R
\name{logRatioToFoldChange}
\alias{logRatioToFoldChange}
\title{Log ratio to fold change}
\usage{
logRatioToFoldChange(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
Modified object, with log ratios converted to fold changes.
}
\description{
Log ratio to fold change
}
\examples{
class(logRatioToFoldChange)
}
