% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeLabel.R
\name{makeLabel}
\alias{makeLabel}
\title{Make a plot axis or legend label out of a character string}
\usage{
makeLabel(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{character(1)}.
Lowercase string.
}
\description{
Doesn't modify strings already containing a space or multi-letter acronym.
}
\examples{
class(makeLabel)
}
\seealso{
\itemize{
\item \code{\link[=makeWords]{makeWords()}}.
\item \code{\link[stringr:case]{stringr::str_to_lower()}}
}
}
