% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeNames.R
\name{makeNames}
\alias{makeNames}
\title{Make syntactically valid names}
\usage{
makeNames(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
Modified object.
}
\description{
For \code{atomic} vectors, these functions will sanitize the values. Otherwise,
they will set \code{\link[base:names]{names()}}, \code{\link[base:rownames]{rownames()}},
and/or \code{\link[base:colnames]{colnames()}} without modification of the values.
}
\details{
Sanitizes names using underscores instead of dots, the convention used by
\code{\link[base:make.names]{make.names()}}.
}
\examples{
class(makeNames)
}
\seealso{
\code{\link[base:make.names]{make.names()}}.
}
