% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeSampleData.R
\name{makeSampleData}
\alias{makeSampleData}
\title{Make sample data}
\usage{
makeSampleData(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{DataFrame}.
}
\description{
Make sample data
}
\examples{
showMethods("makeSampleData")
}
