% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeTitle.R
\name{makeTitle}
\alias{makeTitle}
\title{Make a title out of character string}
\usage{
makeTitle(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{character(1)}.
British style sentence case.
}
\description{
Doesn't modify strings already containing a space or multi-letter acronym.
}
\examples{
class(makeTitle)
}
\seealso{
\itemize{
\item \code{\link[=makeWords]{makeWords()}}.
\item \code{\link[=sentenceCase]{sentenceCase()}}.
\item \code{\link[stringr:str_to_sentence]{stringr::str_to_sentence()}}.
}
}
