% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeWords.R
\name{makeWords}
\alias{makeWords}
\title{Convert syntactic names to words separated by spaces}
\usage{
makeWords(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{character}.
Words separated by spaces.
}
\description{
Convert syntactic names to words separated by spaces
}
\examples{
showMethods("makeWords")
}
