% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapCells.R
\name{mapCells}
\alias{mapCells}
\title{Map input to cell identifiers}
\usage{
mapCells(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{character}.
}
\description{
Map input to cell identifiers
}
\examples{
showMethods("mapCells")
}
