% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapColnames.R
\name{mapColnames}
\alias{mapColnames}
\title{Map input to column names}
\usage{
mapColnames(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{character}.
}
\description{
Map input to column names
}
\examples{
class(mapColnames)
}
