% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapGenes.R
\name{mapGenes}
\alias{mapGenes}
\title{Map input to gene identifiers}
\usage{
mapGenes(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{character}.
}
\description{
Map input to gene identifiers
}
\examples{
class(mapGenes)
}
