% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapGenesToIds.R
\name{mapGenesToIds}
\alias{mapGenesToIds}
\title{Map genes (gene names) to gene identifiers}
\usage{
mapGenesToIds(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{character}.
}
\description{
Map genes (gene names) to gene identifiers
}
\examples{
showMethods("mapGenesToIds")
}
