% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapGenesToRownames.R
\name{mapGenesToRownames}
\alias{mapGenesToRownames}
\title{Map genes to row names}
\usage{
mapGenesToRownames(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\description{
Map genes to row names
}
