% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapGenesToSymbols.R
\name{mapGenesToSymbols}
\alias{mapGenesToSymbols}
\title{Map genes (gene identifiers) to gene names (symbols)}
\usage{
mapGenesToSymbols(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{character}.
}
\description{
Map genes (gene identifiers) to gene names (symbols)
}
