% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapRownames.R
\name{mapRownames}
\alias{mapRownames}
\title{Map input to row names}
\usage{
mapRownames(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{character}.
}
\description{
Map input to row names
}
\examples{
showMethods("mapRownames")
}
