% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapToDataFrame.R
\name{mapToDataFrame}
\alias{mapToDataFrame}
\title{Map elements to a data frame by row binding}
\usage{
mapToDataFrame(x, ...)
}
\arguments{
\item{x}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{DataFrame}.
}
\description{
Map elements to a data frame by row binding
}
\examples{
showMethods("mapToDataFrame")
}
\seealso{
\itemize{
\item \code{purrr::map_dfr()}.
\item \code{data.table::rbindlist()}.
\item \code{rbindToDataFrame()}.
}
}
