% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markdown.R
\name{markdown}
\alias{markdown}
\title{Markdown}
\usage{
markdown(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
Markdown output.
}
\description{
Markdown
}
\examples{
class(markdown)
}
