% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markdownPlots.R
\name{markdownPlots}
\alias{markdownPlots}
\title{Markdown plots}
\usage{
markdownPlots(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
Graphical output of plots.
Separated by Markdown headers.
}
\description{
Markdown plots
}
\examples{
showMethods("markdownPlots")
}
