% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markdownTables.R
\name{markdownTables}
\alias{markdownTables}
\alias{topTables}
\title{Markdown tables}
\usage{
markdownTables(object, ...)

topTables(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{kable}.
Markdown tables.
}
\description{
Markdown tables
}
\examples{
showMethods("markdownTables")
}
