% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchAll.R
\name{matchAll}
\alias{matchAll}
\title{Match all positions of an argument}
\usage{
matchAll(x, table, ...)
}
\arguments{
\item{x}{The values to be matched.}

\item{table}{The values to be matched against.}

\item{...}{Additional arguments.}
}
\value{
\code{list}.
A list of positional vectors corresponding to values defined in \code{table} the
same size as \code{x}.
}
\description{
Match all positions of an argument
}
\details{
\code{matchAll()} behaves like base \code{match()}, but is intended to return a list
of all position matches of the argument defined in \code{x}.
}
\examples{
showMethods("matchAll")
}
