% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchCellColumn.R
\name{matchCellColumn}
\alias{matchCellColumn}
\title{Match cell identifier column}
\usage{
matchCellColumn(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{character(1)}.
}
\description{
Match cell identifier column
}
\examples{
showMethods("matchCellColumn")
}
