% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchGeneColumn.R
\name{matchGeneColumn}
\alias{matchGeneColumn}
\title{Match gene identifier column}
\usage{
matchGeneColumn(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{character(1)}.
}
\description{
Match gene identifier column
}
\examples{
showMethods("matchGeneColumn")
}
