% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchNested.R
\name{matchNested}
\alias{matchNested}
\title{Match inside nested elements}
\usage{
matchNested(x, table, ...)
}
\arguments{
\item{x}{The values to be matched.}

\item{table}{The values to be matched against.
Applies across rows for \code{DataFrame} method.}

\item{...}{Additional arguments.}
}
\value{
\code{integer}.
A positional vector corresponding to values defined in \code{table} the same
size as \code{x}.
}
\description{
Match inside nested elements
}
\details{
Intentionally only performs exact matching. Refer to \code{filterNested} function
for partial matching support with regular expressions.
}
\examples{
showMethods("matchNested")
}
