% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchSampleColumn.R
\name{matchSampleColumn}
\alias{matchSampleColumn}
\title{Match sample identifier column}
\usage{
matchSampleColumn(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{character(1)}.
}
\description{
Match sample identifier column
}
\examples{
showMethods("matchSampleColumn")
}
