% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchTranscriptColumn.R
\name{matchTranscriptColumn}
\alias{matchTranscriptColumn}
\title{Match transcript identifier column}
\usage{
matchTranscriptColumn(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{character(1)}.
}
\description{
Match transcript identifier column
}
\examples{
showMethods("matchTranscriptColumn")
}
