% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcolnames.R
\name{mcolnames}
\alias{mcolnames}
\alias{mcolnames<-}
\title{Metadata column names}
\usage{
mcolnames(x, ...)

mcolnames(x, ...) <- value
}
\arguments{
\item{x}{Object.}

\item{...}{Additional arguments.}

\item{value}{Value.}
}
\description{
Metadata column names
}
\seealso{
\code{S4Vectors::mcols()}.
}
