% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/melt.R
\name{melt}
\alias{melt}
\title{Melt data from wide to long format}
\usage{
melt(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{DataFrame}.
}
\description{
Unpivot column data from wide format to long format.
}
\examples{
showMethods("melt")
}
\seealso{
\itemize{
\item https://seananderson.ca/2013/10/19/reshape/
\item \code{reshape2::melt} (deprecated).
\item \code{tidyr::pivot_longer}.
\item Python \code{pandas.melt}.
}
}
