% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meltCounts.R
\name{meltCounts}
\alias{meltCounts}
\title{Melt count matrix into long format}
\usage{
meltCounts(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{DataFrame}.
Melted data frame containing these columns:
\itemize{
\item \code{colname} (e.g. sample ID).
\item \code{rowname} (e.g. gene ID).
\item \code{counts}.
}
}
\description{
Melt count matrix into long format
}
\examples{
class(meltCounts)
}
\seealso{
\itemize{
\item \code{gather()}.
\item \code{reshape2::melt()}, deprecated in favor of tidyr.
}
}
