% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata2.R
\name{metadata2}
\alias{metadata2}
\alias{metadata2<-}
\title{Metadata}
\usage{
metadata2(x, which, ...)

metadata2(x, which, ...) <- value
}
\arguments{
\item{x}{Object.}

\item{...}{Additional arguments.}

\item{value}{Value to assign.}
}
\value{
\code{list}.
}
\description{
Metadata
}
\examples{
class(metadata2)
}
