% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metricsPerSample.R
\name{metricsPerSample}
\alias{metricsPerSample}
\title{Quality control metrics per sample}
\usage{
metricsPerSample(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\description{
Quality control metrics per sample
}
