% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{missingOrNULL-class}
\alias{missingOrNULL-class}
\title{Missing or NULL class union}
\description{
Missing or NULL class union
}
\note{
Updated 2021-02-02.
}
\examples{
showClass("missingOrNULL")
}
