% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutations.R
\name{mutations}
\alias{mutations}
\title{Mutations}
\usage{
mutations(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{CharacterList}.
}
\description{
Mutations
}
\examples{
showMethods("mutations")
}
