% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ncbi2Ensembl.R
\name{Ncbi2Ensembl}
\alias{Ncbi2Ensembl}
\alias{Entrez2Ensembl}
\title{NCBI-to-Ensembl gene identifier mappings}
\usage{
Ncbi2Ensembl(object, ...)

Entrez2Ensembl(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{Ncbi2Ensembl}.
}
\description{
NCBI-to-Ensembl gene identifier mappings
}
\examples{
showMethods("Ncbi2Ensembl")
}
