% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NcbiToEnsembl.R
\name{NcbiToEnsembl}
\alias{NcbiToEnsembl}
\title{NCBI-to-Ensembl gene identifier mappings}
\usage{
NcbiToEnsembl(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{NcbiToEnsembl}.
}
\description{
NCBI-to-Ensembl gene identifier mappings
}
\examples{
showMethods("NcbiToEnsembl")
}
