% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nest.R
\name{nest}
\alias{nest}
\title{Nest rows of a single column into a list-column}
\usage{
nest(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
Modified object, with a single nested list-column.
}
\description{
Nest rows of a single column into a list-column
}
\examples{
showMethods("nest")
}
\seealso{
\itemize{
\item \code{tidyr::nest()}.
}
}
