% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nesThreshold.R
\name{nesThreshold}
\alias{nesThreshold}
\alias{nesThreshold<-}
\title{Normalized enrichment score (NES) threshold}
\usage{
nesThreshold(object, ...)

nesThreshold(object, ...) <- value
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{value}{Value to assign.}
}
\value{
\code{numeric(1)}.
}
\description{
Normalized enrichment score (NES) threshold
}
\examples{
showMethods("nesThreshold")
}
