% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonzero.R
\name{nonzero}
\alias{nonzero}
\title{Subset object to keep only non-zero rows and columns}
\usage{
nonzero(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
Modified object.
}
\description{
Subset object to keep only non-zero rows and columns
}
\examples{
class(nonzero)
}
