% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonzeroRowsAndCols.R
\name{nonzeroRowsAndCols}
\alias{nonzeroRowsAndCols}
\title{Subset object to keep only non-zero rows and columns}
\usage{
nonzeroRowsAndCols(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
Modified object.
}
\description{
Subset object to keep only non-zero rows and columns
}
\examples{
class(nonzeroRowsAndCols)
}
