% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioc-normalize.R
\name{normalize}
\alias{normalize}
\title{Normalize an object}
\usage{
normalize(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
Modified object.
}
\description{
Normalize an object
}
\examples{
showMethods("normalize")
}
