% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioconductor-02-generics.R
\name{organism}
\alias{organism}
\alias{organism<-}
\title{Organism}
\usage{
organism(object)
organism(object) <- value
}
\arguments{
\item{object}{Object.}

\item{value}{Value to assign.}
}
\value{
\code{character(1)}.
}
\description{
Get or set the organism (i.e. species) of an object.
}
\examples{
showMethods("organism")
}
