% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{params}
\alias{params}
\title{Parameters}
\arguments{
\item{object}{Object.}

\item{value}{Value.}

\item{x}{Object.}

\item{y}{Secondary object.}

\item{...}{Additional arguments.}
}
\description{
Parameters
}
\keyword{internal}
