% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBarcodeRanks.R
\name{plotBarcodeRanks}
\alias{plotBarcodeRanks}
\title{Plot barcode ranks}
\usage{
plotBarcodeRanks(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{ggplot} grid.
}
\description{
Plot barcode ranks
}
\examples{
class(plotBarcodeRanks)
}
