% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBaseMean.R
\name{plotBaseMean}
\alias{plotBaseMean}
\title{Plot base mean distribution}
\usage{
plotBaseMean(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{ggplot}.
}
\description{
The base mean is the mean of normalized counts of all samples, normalizing
for sequencing depth.
}
\examples{
showMethods("plotBaseMean")
}
