% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCellCountsPerCluster.R
\name{plotCellCountsPerCluster}
\alias{plotCellCountsPerCluster}
\title{Plot cell counts per cluster}
\usage{
plotCellCountsPerCluster(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\description{
Plot cell counts per cluster
}
