% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCleveland.R
\name{plotCleveland}
\alias{plotCleveland}
\title{Cleveland plot}
\usage{
plotCleveland(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{ggplot}.
}
\description{
Cleveland plot
}
\examples{
showMethods("plotCleveland")
}
