% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotContrastScatter.R
\name{plotContrastScatter}
\alias{plotContrastScatter}
\title{Plot scatterplot of differential expression contrast}
\usage{
plotContrastScatter(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{ggplot}.
}
\description{
Plot scatterplot of differential expression contrast
}
\examples{
showMethods("plotContrastScatter")
}
