% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCorrelation.R
\name{plotCorrelation}
\alias{plotCorrelation}
\title{Correlation X-Y scatterplot}
\usage{
plotCorrelation(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
Plot.
}
\description{
Correlation X-Y scatterplot
}
\examples{
showMethods("plotCorrelation")
}
