% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCorrelationHeatmap.R
\name{plotCorrelationHeatmap}
\alias{plotCorrelationHeatmap}
\title{Correlation heatmap}
\usage{
plotCorrelationHeatmap(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\description{
Correlation heatmap
}
