% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCountDensity.R
\name{plotCountDensity}
\alias{plotCountDensity}
\title{Plot count density}
\usage{
plotCountDensity(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{ggplot}.
}
\description{
Generally, we expect similar count spreads for all genes between samples
unless the total expressed RNA per sample is different.
}
\examples{
class(plotCountDensity)
}
