% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCountsCorrelation.R
\name{plotCountsCorrelation}
\alias{plotCountsCorrelation}
\title{Plot counts correlation}
\usage{
plotCountsCorrelation(x, y, ...)
}
\arguments{
\item{x}{Object.}

\item{y}{Secondary object.}

\item{...}{Additional arguments.}
}
\value{
Graphical output.
}
\description{
Compare gene expression correlation across multiple data sets.
}
\examples{
class(plotCountsCorrelation)
}
