% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCountsCorrelationHeatmap.R
\name{plotCountsCorrelationHeatmap}
\alias{plotCountsCorrelationHeatmap}
\title{Plot counts correlation heatmap}
\usage{
plotCountsCorrelationHeatmap(x, y, ...)
}
\arguments{
\item{x}{Object.}

\item{y}{Object.}

\item{...}{Additional arguments.}
}
\value{
Graphical output.
}
\description{
Compare feature expression correlation across two data sets.
}
\examples{
showMethods("plotCountsCorrelationHeatmap")
}
