% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCountsPerBiotype.R
\name{plotCountsPerBiotype}
\alias{plotCountsPerBiotype}
\title{Plot counts per biotype}
\usage{
plotCountsPerBiotype(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{ggplot}.
}
\description{
Plot counts per biotype
}
\examples{
showMethods("plotCountsPerBiotype")
}
