% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCountsPerBroadClass.R
\name{plotCountsPerBroadClass}
\alias{plotCountsPerBroadClass}
\title{Plot counts per broad class definition}
\usage{
plotCountsPerBroadClass(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
Plot.
}
\description{
Plot counts per broad class definition
}
\examples{
showMethods("plotCountsPerBroadClass")
}
