% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCountsPerCell.R
\name{plotCountsPerCell}
\alias{plotCountsPerCell}
\title{Plot counts per cell}
\usage{
plotCountsPerCell(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{ggplot}.
}
\description{
Plot the disambiguated counts per cell.
}
\details{
"Counts" refer to universal molecular identifier (UMI) counts for
droplet-based scRNA-seq data.
}
\examples{
showMethods("plotCountsPerCell")
}
\seealso{
\code{\link[=plotCountsVsFeatures]{plotCountsVsFeatures()}}.
}
