% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCountsPerFeature.R
\name{plotCountsPerFeature}
\alias{plotCountsPerFeature}
\title{Plot counts per feature}
\usage{
plotCountsPerFeature(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{ggplot}.
}
\description{
Generally, we expect similar count spreads for all genes between samples
unless the library sizes or total RNA expression are different.
}
\examples{
showMethods("plotCountsPerFeature")
}
