% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDegPca.R
\name{plotDegPca}
\alias{plotDegPca}
\title{Plot differentially expressed gene principal component analysis}
\usage{
plotDegPca(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
Plot.
}
\description{
This function is an extension of \code{plotPCA()} that is optimized for automatic
handling of differentially expressed genes, rather than requiring manual
input of a gene vector or subset object.
}
\details{
To adjust the annotation columns, modify the \code{colData()} of the \code{counts}
argument, which must contain/extend a \code{SummarizedExperiment}.
}
\examples{
showMethods("plotDegPca")
}
