% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDegStackedBar.R
\name{plotDegStackedBar}
\alias{plotDegStackedBar}
\title{Stacked bar plot of DEGs}
\usage{
plotDegStackedBar(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{ggplot}.
}
\description{
Stacked bar plot of DEGs
}
\examples{
showMethods("plotDegStackedBar")
}
