% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDegUpset.R
\name{plotDegUpset}
\alias{plotDegUpset}
\title{UpSet plot of directional DEG intersections across contrasts}
\usage{
plotDegUpset(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
Plot.
}
\description{
UpSet plot of directional DEG intersections across contrasts
}
\examples{
showMethods("plotDegUpset")
}
