% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioconductor-02-generics.R
\name{plotDispEsts}
\alias{plotDispEsts}
\title{Plot dispersion estimates}
\usage{
plotDispEsts(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{ggplot}.
}
\description{
Plot dispersion estimates
}
\examples{
showMethods("plotDispEsts")
}
