% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDot.R
\name{plotDot}
\alias{plotDot}
\title{Dot plot}
\usage{
plotDot(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{ggplot}.
}
\description{
Dot plot
}
\examples{
class(plotDot)
}
