% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotEnrichedGeneSets.R
\name{plotEnrichedGeneSets}
\alias{plotEnrichedGeneSets}
\title{Plot enriched gene sets}
\usage{
plotEnrichedGeneSets(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
Plot.
}
\description{
Plot enriched gene sets
}
\examples{
showMethods("plotEnrichedGeneSets")
}
