% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotEnrichedUpset.R
\name{plotEnrichedUpset}
\alias{plotEnrichedUpset}
\title{UpSet plot of directional enriched pathway intersections across contrasts}
\usage{
plotEnrichedUpset(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
Plot.
}
\description{
UpSet plot of directional enriched pathway intersections across contrasts
}
\examples{
showMethods("plotEnrichedUpset")
}
