% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFeature.R
\name{plotFeature}
\alias{plotFeature}
\title{Plot feature}
\usage{
plotFeature(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{ggplot}.
}
\description{
Plot feature
}
\examples{
showMethods("plotFeature")
}
