% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFeaturesDetected.R
\name{plotFeaturesDetected}
\alias{plotFeaturesDetected}
\title{Plot features detected}
\usage{
plotFeaturesDetected(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{ggplot}.
}
\description{
Visualize the number of features (i.e. genes) detected.
}
\examples{
class(plotFeaturesDetected)
}
