% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFeaturesPerCell.R
\name{plotFeaturesPerCell}
\alias{plotFeaturesPerCell}
\title{Plot features per cell}
\usage{
plotFeaturesPerCell(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{ggplot}.
}
\description{
Plot features per cell
}
\examples{
showMethods("plotFeaturesPerCell")
}
