% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGenderMarkers.R
\name{plotGenderMarkers}
\alias{plotGenderMarkers}
\title{Plot sexually dimorphic gender marker genes}
\usage{
plotGenderMarkers(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\description{
Plot sexually dimorphic gender marker genes
}
