% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGeneEffect.R
\name{plotGeneEffect}
\alias{plotGeneEffect}
\title{Plot gene effect}
\usage{
plotGeneEffect(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{ggplot}.
}
\description{
Plot gene effect
}
\examples{
showMethods("plotGeneEffect")
}
