% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGeneSaturation.R
\name{plotGeneSaturation}
\alias{plotGeneSaturation}
\title{Plot gene detection saturation}
\usage{
plotGeneSaturation(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{ggplot}.
}
\description{
We should observe a linear trend in the number of genes detected with the
number of mapped reads, which indicates that the sample input was not
overloaded.
}
\examples{
class(plotGeneSaturation)
}
