% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGenesDetected.R
\name{plotGenesDetected}
\alias{plotGenesDetected}
\title{Plot genes detected}
\usage{
plotGenesDetected(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\description{
Plot genes detected
}
