% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGeneSet.R
\name{plotGeneSet}
\alias{plotGeneSet}
\title{Plot gene set enrichment}
\usage{
plotGeneSet(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{ggplot}.
}
\description{
Plot gene set enrichment
}
\examples{
class(plotGeneSet)
}
