% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGenesPerCell.R
\name{plotGenesPerCell}
\alias{plotGenesPerCell}
\title{Plot genes per cell}
\usage{
plotGenesPerCell(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{ggplot}.
}
\description{
Plot genes per cell
}
\examples{
class(plotGenesPerCell)
}
