% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotIntergenicMappingRate.R
\name{plotIntergenicMappingRate}
\alias{plotIntergenicMappingRate}
\title{Plot intergenic mapping rate}
\usage{
plotIntergenicMappingRate(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{ggplot}.
}
\description{
Plot intergenic mapping rate
}
