% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotKnownMarkers.R
\name{plotKnownMarkers}
\alias{plotKnownMarkers}
\title{Plot known markers}
\usage{
plotKnownMarkers(object, markers, ...)
}
\arguments{
\item{object}{Object.}

\item{markers}{Object containing gene markers.}

\item{...}{Additional arguments.}
}
\value{
Show graphical output. Invisibly return a \code{ggplot} \code{list}.
}
\description{
Plot known markers
}
