% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotLFC.R
\name{plotLFC}
\alias{plotLFC}
\title{Plot log2 fold change distributions}
\usage{
plotLFC(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{ggplot}.
}
\description{
Plot log2 fold change distributions
}
\examples{
showMethods("plotLFC")
}
