% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMarker.R
\name{plotMarker}
\alias{plotMarker}
\title{Plot cell-type-specific gene markers}
\usage{
plotMarker(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
Show graphical output.
Invisibly return a \code{ggplot} \code{list}.
}
\description{
Visualize gene markers on a reduced dimension plot (e.g. t-SNE, UMAP).
}
\examples{
showMethods("plotMarker")
}
