% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMeanSd.R
\name{plotMeanSd}
\alias{plotMeanSd}
\title{Plot row standard deviations vs. row means}
\usage{
plotMeanSd(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{ggplot}.
}
\description{
Plot row standard deviations vs. row means
}
\examples{
showMethods("plotMeanSd")
}
