% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMitoRatio.R
\name{plotMitoRatio}
\alias{plotMitoRatio}
\title{Plot mitochondrial transcript abundance}
\usage{
plotMitoRatio(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{ggplot}.
}
\description{
Plot mitochondrial transcript abundance
}
