% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMitoVsCoding.R
\name{plotMitoVsCoding}
\alias{plotMitoVsCoding}
\title{Plot mitochondrial vs. coding counts}
\usage{
plotMitoVsCoding(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{ggplot}.
}
\description{
Plot mitochondrial vs. coding counts
}
