% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotNes.R
\name{plotNes}
\alias{plotNes}
\title{Plot normalized enrichment score}
\usage{
plotNes(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{ggplot}.
}
\description{
Plot normalized enrichment score
}
\examples{
showMethods("plotNes")
}
