% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotNovelty.R
\name{plotNovelty}
\alias{plotNovelty}
\title{Plot novelty score}
\usage{
plotNovelty(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{ggplot}.
}
\description{
"Novelty" refers to log10 genes detected per count.
}
\examples{
class(plotNovelty)
}
