% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPca.R
\name{plotPca}
\alias{plotPca}
\title{Principal component analysis plot}
\usage{
plotPca(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
Plot.
}
\description{
Principal component analysis plot
}
\examples{
showMethods("plotPca")
}
\seealso{
\itemize{
\item \code{BiocGenerics::plotPCA}.
}
}
