% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPcaCovariates.R
\name{plotPcaCovariates}
\alias{plotPcaCovariates}
\title{Find correlation between principal components (PCs) and covariates}
\usage{
plotPcaCovariates(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{ggplot}.
}
\description{
Find correlation between principal components (PCs) and covariates
}
\examples{
showMethods("plotPcaCovariates")
}
