% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPcElbow.R
\name{plotPcElbow}
\alias{plotPcElbow}
\title{Plot principal component elbow}
\usage{
plotPcElbow(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{ggplot}.
}
\description{
Calculate the principal component (PC) cutoff using a heuristic approach.
}
\examples{
showMethods("plotPcElbow")
}
