% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPseudoVsAlignedCounts.R
\name{plotPseudoVsAlignedCounts}
\alias{plotPseudoVsAlignedCounts}
\title{Plot pseudo vs. aligned counts}
\usage{
plotPseudoVsAlignedCounts(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{ggplot}.
}
\description{
Plot pseudo vs. aligned counts
}
\examples{
showMethods("plotPseudoVsAlignedCounts")
}
