% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotQc.R
\name{plotQc}
\alias{plotQc}
\title{Quality control}
\usage{
plotQc(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{ggplot}.
}
\description{
Quality control
}
\examples{
showMethods("plotQc")
}
