% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotQuantileHeatmap.R
\name{plotQuantileHeatmap}
\alias{plotQuantileHeatmap}
\title{Quantile heatmap}
\usage{
plotQuantileHeatmap(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
Plot.
}
\description{
Quantile heatmap
}
\examples{
showMethods("plotQuantileHeatmap")
}
