% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotReduction.R
\name{plotReduction}
\alias{plotReduction}
\title{Plot dimensionality reduction}
\usage{
plotReduction(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{ggplot}.
}
\description{
Plot dimensionality reduction
}
\examples{
class(plotReduction)
}
