% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRrnaMappingRate.R
\name{plotRrnaMappingRate}
\alias{plotRrnaMappingRate}
\title{Plot ribosomal RNA (rRNA) mapping rate}
\usage{
plotRrnaMappingRate(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{ggplot}.
}
\description{
Clean, high-quality samples should have an rRNA mapping rate below 10\%.
Higher rates are likely indicative of the polyA enrichment or ribo depletion
protocol not having removed all ribosomal RNA (rRNA) transcripts. This will
reduce the number of biologically meaningful reads in the experiment and is
best avoided.
}
\examples{
showMethods("plotRrnaMappingRate")
}
