% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotStackedBarPlot.R
\name{plotStackedBarPlot}
\alias{plotStackedBarPlot}
\title{Plot stacked bar plot}
\usage{
plotStackedBarPlot(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{ggplot}.
}
\description{
Plot stacked bar plot
}
\examples{
class(plotStackedBarPlot)
}
