% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSums.R
\name{plotSums}
\alias{plotSums}
\title{Plot row or column sums}
\usage{
plotSums(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{ggplot}.
}
\description{
Visualize row or column sums using empirical cumulative distribution function
(ECDF) plot.
}
\examples{
showMethods("plotSums")
}
