% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTopGeneEffectPerCell.R
\name{plotTopGeneEffectPerCell}
\alias{plotTopGeneEffectPerCell}
\title{Plot top gene effect per cell}
\usage{
plotTopGeneEffectPerCell(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{ggplot}.
}
\description{
Plot top gene effect per cell
}
\examples{
showMethods("plotTopGeneEffectPerCell")
}
\seealso{
\itemize{
\item DepMapAnalysis package.
}
}
