% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTopGeneEffectPerGroup.R
\name{plotTopGeneEffectPerGroup}
\alias{plotTopGeneEffectPerGroup}
\title{Plot top gene effect per group}
\usage{
plotTopGeneEffectPerGroup(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{ggplot}.
}
\description{
Plot top gene effect per group
}
\examples{
showMethods("plotTopGeneEffectPerGroup")
}
\seealso{
\itemize{
\item DepMapAnalysis package.
}
}
