% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTopMarkers.R
\name{plotTopMarkers}
\alias{plotTopMarkers}
\title{Plot top markers}
\usage{
plotTopMarkers(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
Show graphical output. Invisibly return a \code{ggplot} \code{list}.
}
\description{
Plot top markers
}
