% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTotalCounts.R
\name{plotTotalCounts}
\alias{plotTotalCounts}
\title{Plot total read counts}
\usage{
plotTotalCounts(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\description{
Plot total read counts
}
