% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTsne.R
\name{plotTsne}
\alias{plotTsne}
\title{t-SNE plot}
\usage{
plotTsne(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{ggplot}.
}
\description{
t-SNE plot
}
\examples{
showMethods("plotTsne")
}
