% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotUmap.R
\name{plotUmap}
\alias{plotUmap}
\title{UMAP plot}
\usage{
plotUmap(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{ggplot}.
}
\description{
UMAP plot
}
\examples{
showMethods("plotUmap")
}
