% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCountsVsFeatures.R
\name{plotUMIsVsGenes}
\alias{plotUMIsVsGenes}
\title{Plot count and feature correlation}
\usage{
plotUMIsVsGenes(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{ggplot}.
}
\description{
Plot the disambiguated counts per cell vs. features (i.e. genes or
transcripts) detected.
}
\details{
"Counts" refer to universal molecular identifier (UMI) counts for
droplet-based scRNA-seq data.
}
\examples{
class(plotUMIsVsGenes)
}
\seealso{
\code{\link[=plotCountsPerCell]{plotCountsPerCell()}}.
}
