% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotUpset.R
\name{plotUpset}
\alias{plotUpset}
\title{UpSet plot}
\usage{
plotUpset(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
Plot.
}
\description{
UpSet plot
}
\examples{
showMethods("plotUpset")
}
