% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotVolcano.R
\name{plotVolcano}
\alias{plotVolcano}
\title{Volcano plot}
\usage{
plotVolcano(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{ggplot}.
}
\description{
Volcano plot
}
\note{
Don't apply post hoc alpha level or LFC threshold cutoffs.
}
