% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotWaterfall.R
\name{plotWaterfall}
\alias{plotWaterfall}
\title{Plot waterfall}
\usage{
plotWaterfall(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{ggplot}.
}
\description{
Plot waterfall
}
\examples{
class(plotWaterfall)
}
