% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotZerosVsDepth.R
\name{plotZerosVsDepth}
\alias{plotZerosVsDepth}
\title{Plot zeros vs. depth}
\usage{
plotZerosVsDepth(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\description{
Plot zeros vs. depth
}
