% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictSensitivity.R
\name{predictSensitivity}
\alias{predictSensitivity}
\title{Predict sensitivity}
\usage{
predictSensitivity(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{DFrame}.
}
\description{
Predict sensitivity
}
\examples{
showMethods("predictSensitivity")
}
\seealso{
\itemize{
\item DepMapAnalysis package.
}
}
