% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudobulk.R
\name{pseudobulk}
\alias{pseudobulk}
\title{Pseudobulk}
\usage{
pseudobulk(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
Modified object, with cells aggregated to samples, like a bulk
RNA-seq dataset.
}
\description{
Pseudobulk
}
\note{
Applies to single-cell RNA-seq objects, usually \code{SingleCellExperiment}.
}
\examples{
class(pseudobulk)
}
