% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rankedMatrix.R
\name{rankedMatrix}
\alias{rankedMatrix}
\title{Calculate a ranked matrix}
\usage{
rankedMatrix(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{matrix}.
}
\description{
Calculate a ranked matrix
}
\details{
Ties are resolved automatically by calculating the average.
See the \code{ties.method} parameter in \code{\link[base:rank]{rank()}} for details.
}
\examples{
showMethods("rankedMatrix")
}
\seealso{
\itemize{
\item \code{\link[base:rank]{rank()}}.
\item Not to be confused with \code{Matrix::rankMatrix()}.
}
}
