% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbindToDataFrame.R
\name{rbindToDataFrame}
\alias{rbindToDataFrame}
\title{Row bind elements 1:1 to a data frame}
\usage{
rbindToDataFrame(x, ...)
}
\arguments{
\item{x}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{DataFrame}.
}
\description{
Row bind elements 1:1 to a data frame
}
\examples{
showMethods("rbindToDataFrame")
}
\seealso{
\itemize{
\item \code{mapToDataFrame()}.
}
}
