% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports-bioc.R
\docType{import}
\name{reexports-bioc}
\alias{reexports-bioc}
\alias{AsIs-class}
\alias{\%in\%}
\alias{colSums}
\alias{complete.cases}
\alias{cor}
\alias{do.call}
\alias{expand}
\alias{expand.grid}
\alias{lapply}
\alias{match}
\alias{mcols}
\alias{merge}
\alias{metadata}
\alias{na.omit}
\alias{order}
\alias{rowMeans}
\alias{rowSums}
\alias{setdiff}
\alias{split}
\alias{t}
\alias{table}
\alias{unique}
\alias{unlist}
\title{Bioconductor reexports}
\description{
Bioconductor reexports
}
\keyword{internal}
