% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeNA.R
\name{removeNA}
\alias{removeNA}
\title{Remove rows and columns containing only \code{NA} values}
\usage{
removeNA(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\description{
Remove rows and columns containing only \code{NA} values
}
