% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resultsDiff.R
\name{resultsDiff}
\alias{resultsDiff}
\title{Relative difference of results}
\usage{
resultsDiff(x, y, ...)
}
\arguments{
\item{x}{Object.}

\item{y}{Object.}

\item{...}{Additional arguments.}
}
\value{
Named \code{numeric}.
Names correspond to the features (i.e. genes) analyzed.
}
\description{
Relative difference of results
}
\examples{
showMethods("resultsDiff")
}
