% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resultsMatrix.R
\name{resultsMatrix}
\alias{resultsMatrix}
\title{Results matrix}
\usage{
resultsMatrix(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{matrix}.
}
\description{
Results matrix
}
\examples{
showMethods("resultsMatrix")
}
