% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resultsNames.R
\name{resultsNames}
\alias{resultsNames}
\alias{resultsNames<-}
\title{Results names}
\usage{
resultsNames(object, ...)

resultsNames(object, ...) <- value
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{value}{Value to assign.}
}
\value{
\code{character}.
}
\description{
Results names
}
\examples{
showMethods("resultsNames")
}
\seealso{
\itemize{
\item \code{DESeq2::resultsNames()}.
}
}
