% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resultsTables.R
\name{resultsTables}
\alias{resultsTables}
\title{Results tables}
\usage{
resultsTables(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\description{
Results tables
}
