% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleData.R
\name{sampleData}
\alias{sampleData}
\alias{sampleData<-}
\title{Sample data}
\usage{
sampleData(object, ...)

sampleData(object, ...) <- value
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{value}{Value.}
}
\description{
Sample data
}
