% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleNames.R
\name{sampleNames}
\alias{sampleNames}
\alias{sampleNames<-}
\title{Sample names}
\usage{
sampleNames(object)

sampleNames(object) <- value
}
\arguments{
\item{object}{Object.}

\item{value}{Value to assign.}

\item{...}{Additional arguments.}
}
\description{
Sample names
}
\note{
Updated 2020-12-21.
}
