% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanitizeNA.R
\name{sanitizeNA}
\alias{sanitizeNA}
\title{Sanitize NA values}
\usage{
sanitizeNA(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
Modified object.
Sanitized to contain proper \code{NA} values.
}
\description{
Standardize empty strings (\code{""}), character NAs (\code{"NA"}), and \code{NULL} values
inside a character vector to \code{NA_character_}. Other \code{atomic} data types are
returned unmodified.
}
\examples{
showMethods("sanitizeNA")
}
