% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanitizePercent.R
\name{sanitizePercent}
\alias{sanitizePercent}
\title{Sanitize percentage}
\usage{
sanitizePercent(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\description{
Sanitize percentage
}
