% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectCells.R
\name{selectCells}
\alias{selectCells}
\title{Select cells}
\usage{
selectCells(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Named key value pairs that map to cell line metadata defined in \code{colData}.}
}
\value{
Modified object.
}
\description{
Select cells
}
\examples{
showMethods("selectCells")
}
\seealso{
\itemize{
\item Cellosaurus package.
\item DepMapAnalysis package.
}
}
