% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectSamples.R
\name{selectSamples}
\alias{selectSamples}
\title{Select samples}
\usage{
selectSamples(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\description{
Select samples
}
