% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sem.R
\name{sem}
\alias{sem}
\title{Standard error of the mean}
\usage{
sem(x, ...)
}
\arguments{
\item{x}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{numeric}.
}
\description{
Standard error of the mean
}
\examples{
class(sem)
}
\seealso{
\itemize{
\item https://stackoverflow.com/questions/2676554/
}
}
