% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentenceCase.R
\name{sentenceCase}
\alias{sentenceCase}
\title{Sentence case}
\usage{
sentenceCase(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
Modified object.
}
\description{
Sentence case
}
\examples{
class(sentenceCase)
}
