% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showHeader.R
\name{showHeader}
\alias{showHeader}
\title{Header for object show method}
\usage{
showHeader(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
Console output, via \code{\link[base:cat]{cat()}}.
}
\description{
Header for object show method
}
\examples{
showMethods("showHeader")
}
\seealso{
\code{\link[=show]{show()}}.
}
