% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioconductor-02-generics.R
\name{sizeFactors}
\alias{sizeFactors}
\alias{sizeFactors<-}
\title{Size factors}
\usage{
sizeFactors(object, ...)
sizeFactors(object, ...) <- value
}
\arguments{
\item{object}{Object.}

\item{value}{Value to assign.}

\item{...}{Additional arguments.}
}
\value{
\code{numeric}.
Names correspond to object column names.
}
\description{
Size factors
}
\examples{
showMethods("sizeFactors")
}
\seealso{
\itemize{
\item \code{DESeq2::sizeFactors()}.
\item \code{DESeq2::estimateSizeFactors()}.
\item \code{DESeq2::estimateSizeFactorsForMatrix()}.
}
}
