% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slotAlignedCounts.R
\name{slotAlignedCounts}
\alias{slotAlignedCounts}
\title{Slot aligned counts}
\usage{
slotAlignedCounts(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
Modified object.
}
\description{
Slot aligned counts
}
\examples{
showMethods("slotAlignedCounts")
}
