% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snake.R
\name{snake}
\alias{snake}
\title{Snake case}
\usage{
snake(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
Modified object, with names formatted in snake case.
}
\description{
Format character strings to use snake-style formatting, where word boundaries
are defined by underscores (e.g. \code{this_is_snake_case}).
}
\details{
Snake case is recommended by the tidyverse for variable and function names.
}
\examples{
class(snake)
}
