% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snakeCase.R
\name{snakeCase}
\alias{snakeCase}
\title{Snake case}
\usage{
snakeCase(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
Modified object.
Contains syntatically valid names.
For objects with \code{\link[base:names]{names()}} defined, the underlying data
returns unchanged, except for \code{character} or \code{vector} class.
}
\description{
Format character strings to use snake-style formatting, where word boundaries
are defined by underscores (e.g. \code{this_is_snake_case}).
}
\examples{
class(snakeCase)
}
