% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitByLevel.R
\name{splitByLevel}
\alias{splitByLevel}
\title{Split by factor level}
\usage{
splitByLevel(x, f, ref = FALSE, ...)
}
\arguments{
\item{f}{\code{character(1)}.
Factor column name.}

\item{ref}{\code{logical(1)}.
Include the reference factor level (i.e. first defined) in each split.
Useful for pairwise contrasts.}
}
\value{
\code{list}
List of data frames split by factors.
}
\description{
Split by factor level
}
\note{
Updated 2020-09-14.
}
\examples{
class(splitByLevel)
}
\seealso{
\itemize{
\item \code{split()}
\item \code{S4Vectors::split()}
\item \code{dplyr::group_split()}
}
}
