% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spread.R
\name{spread}
\alias{spread}
\title{Spread a key-value pair across multiple columns}
\usage{
spread(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{DataFrame}.
}
\description{
Spread a key-value pair across multiple columns
}
\examples{
class(spread)
}
\seealso{
\code{tidyr::spread()}.
}
