% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardizeCells.R
\name{standardizeCells}
\alias{standardizeCells}
\title{Standardize cell line names}
\usage{
standardizeCells(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{character}.
}
\description{
Standardize cell line names
}
\examples{
showMethods("standardizeCells")
}
