% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stripExonVersions.R
\name{stripExonVersions}
\alias{stripExonVersions}
\title{Strip exon identifier versions}
\usage{
stripExonVersions(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
Modified object.
Contains gene identifiers without version numbers.
}
\description{
Strip exon identifier versions
}
\details{
Intentionally does not strip transcript versions, if defined in the object.
To perform that action in parallel, use \code{stripVersions} instead.
}
\examples{
showMethods("stripExonVersions")
}
