% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stripTranscriptVersions.R
\name{stripTranscriptVersions}
\alias{stripTranscriptVersions}
\title{Strip transcript identifier versions}
\usage{
stripTranscriptVersions(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
Modified object.
Contains transcript identifiers without version numbers.
}
\description{
Strip transcript identifier versions
}
\details{
Intentionally does not strip gene versions, if defined in the object.
To perform that action in parallel, use \code{stripGeneVersions} instead.
}
\examples{
showMethods("stripTranscriptVersions")
}
