% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioconductor-02-generics.R
\name{summary}
\alias{summary}
\title{Summary}
\usage{
summary(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
Console output.
}
\description{
Summary
}
\examples{
showMethods("summary")
}
