% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tail.R
\name{tail}
\alias{tail}
\title{Return the last part of an object}
\usage{
tail(x, ...)
}
\arguments{
\item{x}{Object.}

\item{...}{Additional arguments.}
}
\value{
Console output.
}
\description{
Return the last part of an object
}
\examples{
showMethods("tail")
}
