% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tnbc.R
\name{tnbc}
\alias{tnbc}
\title{Triple-negative breast cancer (TNBC)}
\usage{
tnbc(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{character}.
TNBC patient samples or cell lines.
}
\description{
Triple-negative breast cancer (TNBC)
}
\examples{
showMethods("tnbc")
}
\seealso{
\itemize{
\item Cellosaurus package.
}
}
