% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topCellsPerSample.R
\name{topCellsPerSample}
\alias{topCellsPerSample}
\title{Top cells per sample}
\usage{
topCellsPerSample(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\description{
Top cells per sample
}
