% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topMarkers.R
\name{topMarkers}
\alias{topMarkers}
\title{Top markers}
\usage{
topMarkers(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{grouped_df}.
}
\description{
Top markers
}
