% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topTables.R
\name{topTables}
\alias{topTables}
\title{Top tables}
\usage{
topTables(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{kable}.
Markdown tables.
}
\description{
Top tables of differentially expressed genes.
}
\examples{
showMethods("topTables")
}
