% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpm.R
\name{tpm}
\alias{tpm}
\title{Transcripts per million}
\usage{
tpm(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{matrix}
}
\description{
Transcripts per million
}
\examples{
showMethods("tpm")
}
