% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcriptIds.R
\name{transcriptIds}
\alias{transcriptIds}
\title{Transcript identifiers}
\usage{
transcriptIds(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{character}.
}
\description{
Convenience function for quickly returning the stable transcript identifiers.
}
\examples{
showMethods("transcriptIds")
}
