% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcriptNames.R
\name{transcriptNames}
\alias{transcriptNames}
\title{Transcript names}
\usage{
transcriptNames(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{character}.
}
\description{
Convenience function for quickly returning transcript names that map to the
stable transcript identifiers.
}
\examples{
showMethods("transcriptNames")
}
