% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformType.R
\name{transformType}
\alias{transformType}
\title{Variance-stabilizing transformation type}
\usage{
transformType(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{character(1)}.
Transformation method name.
}
\description{
Variance-stabilizing transformation type
}
\examples{
class(transformType)
}
