% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tx2Gene.R
\name{Tx2Gene}
\alias{Tx2Gene}
\title{Transcript-to-gene mappings}
\usage{
Tx2Gene(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{Tx2Gene}.
}
\description{
Transcript-to-gene mappings
}
\examples{
showMethods("Tx2Gene")
}
