% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TxToGene.R
\name{TxToGene}
\alias{TxToGene}
\title{Transcript-to-gene mappings}
\usage{
TxToGene(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{TxToGene}.
}
\description{
Transcript-to-gene mappings
}
\examples{
showMethods("TxToGene")
}
