% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniteInterestingGroups.R
\name{uniteInterestingGroups}
\alias{uniteInterestingGroups}
\title{Unite interesting groups into a single column}
\usage{
uniteInterestingGroups(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
Modified object.
Now contains an \code{interestingGroups} column.
}
\description{
Create a single interesting groups column (\code{interestingGroups}) used for
coloring in plots. When multiple interesting groups are present, unite into a
single column, delimited by a colon.
}
\examples{
showMethods("uniteInterestingGroups")
}
