% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unlist2.R
\name{unlist2}
\alias{unlist2}
\title{Flatten lists}
\usage{
unlist2(x, ...)
}
\arguments{
\item{x}{Object.}

\item{...}{Additional arguments.}
}
\value{
Unlisted object derived from original input.
For example, \code{DFrame} derived from \code{DFrameList}.
}
\description{
Flatten lists
}
\details{
Intended primarily for complex S4 classes, such as \code{DFrameList}.
Intentionally does not support recursion.
}
\examples{
showMethods("unlist2")
}
\seealso{
\itemize{
\item \code{unlist()}.
}
}
