% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unnest.R
\name{unnest}
\alias{unnest}
\title{Unnest a single list-column into expanded rows}
\usage{
unnest(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
Modified object, with a single unnested list-column expanded into
rows.
}
\description{
Unnest a single list-column into expanded rows
}
\examples{
showMethods("unnest")
}
\seealso{
\itemize{
\item \code{tidyr::unnest()}.
}
}
