% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioconductor-02-generics.R
\name{updateObject}
\alias{updateObject}
\title{Update object}
\usage{
updateObject(object, ..., verbose = FALSE)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{verbose}{\code{logical(1)}.
Verbose output.}
}
\value{
Modified object.
}
\description{
Update object
}
\examples{
showMethods("updateObject")
}
