% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upperCamel.R
\name{upperCamel}
\alias{upperCamel}
\title{Upper camel case}
\usage{
upperCamel(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
Modified object, with names formatted in upper camel case.
}
\description{
Format character strings to use upper camel-style formatting, where word
boundaries are defined by capitlization only (e.g. \code{ThisIsCamelCase}).
}
\details{
Note that lower camel case is generally recommended in R over the use of
upper camel case. However, upper camel case is recommended by Bioconductor
for S4 class names and corresponding generators, but not variables or
functions.
}
\examples{
class(upperCamel)
}
