% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upsetMatrix.R
\name{upsetMatrix}
\alias{upsetMatrix}
\title{UpSet matrix}
\usage{
upsetMatrix(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\verb{logical matrix}.
}
\description{
UpSet matrix
}
\examples{
showMethods("upsetMatrix")
}
\seealso{
\itemize{
\item \code{UpSetR::fromList()}.
\item \code{intersect()}, \code{setdiff()}.
}
}
