% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zerosVsDepth.R
\name{zerosVsDepth}
\alias{zerosVsDepth}
\title{Percentage of zeros vs. library depth}
\usage{
zerosVsDepth(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\description{
Percentage of zeros vs. library depth
}
