% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zscore.R
\name{zscore}
\alias{zscore}
\title{Calculate Z-score}
\usage{
zscore(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
Numeric vector or matrix.
}
\description{
Calculate Z-score
}
\examples{
showMethods("zscore")
}
\seealso{
\itemize{
\item \code{scale()}.
\item \code{limma::zscore()}.
}
}
