% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{AcidGenerics-package}
\alias{AcidGenerics}
\alias{AcidGenerics-package}
\title{AcidGenerics}
\description{
S4 generic functions for Acid Genomics packages.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://acidgenerics.acidgenomics.com/}
  \item \url{https://github.com/acidgenomics/AcidGenerics/}
  \item Report bugs at \url{https://github.com/acidgenomics/AcidGenerics/issues/}
}

}
\author{
\strong{Maintainer}: Michael Steinbaugh \email{mike@steinbaugh.com} (\href{https://orcid.org/0000-0002-2403-2221}{ORCID})

Other contributors:
\itemize{
  \item Acid Genomics [copyright holder, funder]
}

}
\keyword{internal}
