% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregateCellsToSamples.R
\name{aggregateCellsToSamples}
\alias{aggregateCellsToSamples}
\title{Aggregate cells to samples}
\usage{
aggregateCellsToSamples(x, ...)
}
\arguments{
\item{x}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{SummarizedExperiment}.
Object with cell-level counts aggregated to sample-level.
}
\description{
Utilty function that factilites cell-to-sample aggregation. By default, this
function will sum the counts across cells to sample level.
}
\note{
This function is intended primarily for quality control analysis.
}
\examples{
class(aggregateCellsToSamples)
}
