% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregateCols.R
\name{aggregateCols}
\alias{aggregateCols}
\title{Aggregate columns}
\usage{
aggregateCols(x, ...)
}
\arguments{
\item{x}{Object.}

\item{...}{Additional arguments.}
}
\value{
Modified object of same class, with columns aggregated.
}
\description{
Aggregate columns
}
\examples{
class(aggregateCols)
}
