% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregateRows.R
\name{aggregateRows}
\alias{aggregateRows}
\title{Aggregate rows}
\usage{
aggregateRows(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
Modified object of same class, with rows aggregated.
}
\description{
Aggregate rows
}
\examples{
class(aggregateRows)
}
