% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alphaThreshold.R
\name{alphaThreshold}
\alias{alphaThreshold}
\alias{alphaThreshold<-}
\title{Alpha threshold}
\usage{
alphaThreshold(object, ...)

alphaThreshold(object, ...) <- value
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{value}{Value to assign.}
}
\value{
\code{numeric(1)}.
Alpha cutoff level.
}
\description{
Alpha threshold
}
\examples{
class(alphaThreshold)
}
