% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.DataFrame.R
\name{as.DataFrame}
\alias{as.DataFrame}
\title{Coerce object to S4 DataFrame}
\usage{
as.DataFrame(x, ...)
}
\arguments{
\item{x}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{DataFrame}.
}
\description{
Coerce object to S4 DataFrame
}
\examples{
showMethods("as.DataFrame")
}
