% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioconductor-02-generics.R
\name{as.list}
\alias{as.list}
\title{Coerce object to a list}
\usage{
as.list(x, ...)
}
\arguments{
\item{x}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{list}.
}
\description{
Coerce object to a list
}
\examples{
showMethods("as.list")
}
