% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{bioverbs-package}
\alias{bioverbs}
\alias{bioverbs-package}
\title{bioverbs}
\description{
S4 generic functions for bioinformatics.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://bioverbs.acidgenomics.com/}
  \item \url{https://github.com/acidgenomics/bioverbs/}
  \item Report bugs at \url{https://github.com/acidgenomics/bioverbs/issues/}
}

}
\author{
\strong{Maintainer}: Michael Steinbaugh \email{mike@steinbaugh.com} (\href{https://orcid.org/0000-0002-2403-2221}{ORCID})

Other contributors:
\itemize{
  \item Acid Genomics [copyright holder, funder]
}

}
\keyword{internal}
