% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camelCase.R
\name{camelCase}
\alias{camelCase}
\title{Camel case}
\usage{
camelCase(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
Modified object.
Contains syntatically valid names.
For objects with \code{\link[base:names]{names()}} defined, the underlying data
returns unchanged, except for \code{character} or \code{vector} class.
}
\description{
Format character strings to use (lower) camel-style formatting, where word
boundaries are defined by capitlization only (e.g. \code{thisIsCamelCase}).
}
\details{
Camel case is recommended by Bioconductor for variable and function names.
}
\examples{
showMethods("camelCase")
}
