% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellCountsPerCluster.R
\name{cellCountsPerCluster}
\alias{cellCountsPerCluster}
\title{Cell counts per cluster}
\usage{
cellCountsPerCluster(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{DataFrame}.
}
\description{
Cell counts per cluster
}
\examples{
class(cellCountsPerCluster)
}
