% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterCellCountsPerSample.R
\name{clusterCellCountsPerSample}
\alias{clusterCellCountsPerSample}
\title{Cluster cell counts per sample}
\usage{
clusterCellCountsPerSample(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{grouped_df}.
Grouped by \code{sampleName} column, arranged by abundance.
}
\description{
Cluster cell counts per sample
}
