% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapseToString.R
\name{collapseToString}
\alias{collapseToString}
\title{Collapse to string}
\usage{
collapseToString(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\description{
Collapse to string
}
