% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrastNames.R
\name{contrastNames}
\alias{contrastNames}
\alias{contrastNames<-}
\title{Contrast names}
\usage{
contrastNames(object, ...)

contrastNames(object, ...) <- value
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{value}{Value to assign.}
}
\value{
\code{character}.
}
\description{
Contrast names
}
\examples{
showMethods("contrastNames")
}
