% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertGenesToSymbols.R
\name{convertGenesToSymbols}
\alias{convertGenesToSymbols}
\title{Convert genes to symbols}
\usage{
convertGenesToSymbols(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\description{
Convert genes to symbols
}
