% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertSampleIdsToNames.R
\name{convertSampleIdsToNames}
\alias{convertSampleIdsToNames}
\title{Convert sample identifiers to names}
\usage{
convertSampleIdsToNames(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
Modified object of same class, with valid sample IDs remapped to
human-readable sample names.
}
\description{
Convert sample identifiers to names
}
\examples{
showMethods("convertSampleIdsToNames")
}
