% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertSymbolsToGenes.R
\name{convertSymbolsToGenes}
\alias{convertSymbolsToGenes}
\title{Convert symbols to genes}
\usage{
convertSymbolsToGenes(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
Modified object of same class, with human-readable gene symbols
remapped back to stable gene identifiers.
}
\description{
Convert symbols to genes
}
\examples{
showMethods("convertSymbolsToGenes")
}
