% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts.R
\name{counts}
\alias{counts}
\title{Counts}
\usage{
counts(object, ...)
counts(object) <- value
}
\arguments{
\item{object}{Object.}

\item{value}{Value to assign.}

\item{...}{Additional arguments.}
}
\value{
\code{matrix}.
}
\description{
Counts
}
\examples{
showMethods("counts")
}
