% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{defunct}
\alias{defunct}
\alias{plotGenesPerCell}
\alias{plotGenesDetected}
\alias{plotUMIsPerCell}
\alias{plotUMIsVsGenes}
\alias{meltCounts}
\alias{clusterCellCountsPerSample}
\alias{flatFiles}
\alias{plotGSEATable}
\alias{plotDot}
\alias{matchRowNameColumn}
\alias{markdown}
\alias{mcolnames}
\alias{mcolnames<-}
\title{Defunct functions}
\usage{
plotGenesPerCell(object, ...)

plotGenesDetected(object, ...)

plotUMIsPerCell(object, ...)

plotUMIsVsGenes(object, ...)

meltCounts(object, ...)

clusterCellCountsPerSample(object, ...)

flatFiles(object, ...)

plotGSEATable(object, ...)

plotDot(object, ...)

matchRowNameColumn(object, ...)

markdown(object, ...)

mcolnames(x, ...)

mcolnames(x, ...) <- value
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{x}{Object.}

\item{value}{Value to assign.}
}
\value{
\code{.Defunct}.
}
\description{
Defunct functions
}
\examples{
a <- function(...) {
    .Defunct("b")
}

withCallingHandlers(
    expr = tryCatch(
        expr = a(1L),
        error = function(e) {
            print(e)
            invisible()
        }
    )
)
}
\seealso{
\link[base:Defunct]{Defunct}.
}
\keyword{internal}
