% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{deprecated}
\alias{deprecated}
\alias{plotGenesPerCell}
\alias{plotGenesDetected}
\alias{plotUMIsPerCell}
\alias{plotUMIsVsGenes}
\alias{meltCounts}
\alias{clusterCellCountsPerSample}
\alias{flatFiles}
\alias{plotGSEATable}
\alias{plotDot}
\alias{matchRowNameColumn}
\alias{markdown}
\alias{mcolnames}
\alias{mcolnames<-}
\title{Deprecated functions}
\usage{
plotGenesPerCell(object, ...)

plotGenesDetected(object, ...)

plotUMIsPerCell(object, ...)

plotUMIsVsGenes(object, ...)

meltCounts(object, ...)

clusterCellCountsPerSample(object, ...)

flatFiles(object, ...)

plotGSEATable(object, ...)

plotDot(object, ...)

matchRowNameColumn(object, ...)

markdown(object, ...)

mcolnames(x, ...)

mcolnames(x, ...) <- value
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{x}{Object.}

\item{value}{Value to assign.}
}
\value{
\code{.Deprecated}.
}
\description{
Deprecated functions
}
\examples{
a <- function(...) {
    .Deprecated("b")
    b(...)
}

b <- function(x) {
    x + 1L
}

withCallingHandlers(
    expr = a(1L),
    warning = function(w) {
        print(w)
        invokeRestart("muffleWarning")
    }
)
}
\seealso{
\link[base:Deprecated]{Deprecated}.
}
\keyword{internal}
