% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{deprecated}
\alias{deprecated}
\alias{clusterID}
\alias{plotGenesPerCell}
\alias{plotGenesDetected}
\alias{plotUMIsPerCell}
\alias{plotUMIsVsGenes}
\alias{meltCounts}
\title{Deprecated functions}
\usage{
clusterID(object, ...)

plotGenesPerCell(object, ...)

plotGenesDetected(object, ...)

plotUMIsPerCell(object, ...)

plotUMIsVsGenes(object, ...)

meltCounts(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{.Deprecated}.
}
\description{
Deprecated functions
}
\examples{
a <- function(...) {
    .Deprecated("b")
    b(...)
}

b <- function(x) {
    x + 1L
}

withCallingHandlers(
    expr = a(1L),
    warning = function(w) {
        print(w)
        invokeRestart("muffleWarning")
    }
)
}
\seealso{
\link[base:Deprecated]{Deprecated}.
}
\keyword{internal}
