% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{deprecated}
\alias{deprecated}
\alias{plotGene}
\title{Deprecated functions}
\usage{
plotGene(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{.Deprecated}.
}
\description{
Deprecated functions
}
\keyword{internal}
