% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffExpPerCluster.R
\name{diffExpPerCluster}
\alias{diffExpPerCluster}
\title{Differential expression per cluster}
\usage{
diffExpPerCluster(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
Varies, depending on the method.
}
\description{
Differential expression per cluster
}
\examples{
class(diffExpPerCluster)
}
