% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dottedCase.R
\name{dottedCase}
\alias{dottedCase}
\title{Dotted case}
\usage{
dottedCase(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
Modified object, with names formatted in dotted case.
}
\description{
Dotted case
}
\examples{
class(dottedCase)
}
