% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioc-droplevels.R
\name{droplevels}
\alias{droplevels}
\title{Drop levels}
\usage{
droplevels(x, ...)
}
\arguments{
\item{x}{Object.}

\item{...}{Additional arguments.}
}
\value{
Modified object.
}
\description{
Drop levels
}
\examples{
showMethods("droplevels")
}
