% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{export}
\alias{export}
\title{Export}
\usage{
export(x, ...)
}
\arguments{
\item{x}{Object.}

\item{...}{Additional arguments.}
}
\description{
Export data out of R and write to disk.
}
\details{
This is a wrapper for \code{\link[rio:export]{rio::export()}} that adds support for additional S4
classes in Bioconductor.
}
