% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{export}
\alias{export}
\title{Export}
\usage{
export(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
Invisibily return file paths, either as \code{list} or \code{character}.
}
\description{
Export data out of R and write to disk.
}
\examples{
class(export)
}
