% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneIDs.R
\name{geneIDs}
\alias{geneIDs}
\title{Gene identifiers}
\usage{
geneIDs(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{character}.
}
\description{
Convenience function for quickly returning the stable gene identifiers.
}
\examples{
showMethods("geneIDs")
}
