% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometricMean.R
\name{geometricMean}
\alias{geometricMean}
\title{Geometric mean}
\usage{
geometricMean(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{numeric}.
}
\description{
The geometric mean is the nth root of n products or e to the mean log of \code{x}.
}
\details{
Useful for describing non-normal (i.e. geometric) distributions.
}
\examples{
class(geometricMean)
}
