% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/headtail.R
\name{headtail}
\alias{headtail}
\title{Return the first and last parts of an object}
\usage{
headtail(x, ...)
}
\arguments{
\item{x}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{character}.
}
\description{
Applies to both rows and columns, enabling quick inspection during
interactive use.
}
\details{
Inspired by the print method for \code{DataFrame} class.
}
\examples{
showMethods("headtail")
}
\seealso{
\itemize{
\item \code{utils::head()}, \code{utils::tail()}.
\item \code{cat()}.
}
}
