% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import}
\alias{import}
\title{Import data into R}
\usage{
import(con, ...)
}
\arguments{
\item{con}{File connection.
Typically \code{character(1)} but can also be \code{textConnection}.}

\item{...}{Additional arguments.}
}
\value{
Object.
}
\description{
Import data into R
}
\examples{
showMethods("import")
}
\seealso{
\itemize{
\item \code{BiocIO::import}.
\item \code{rio::import}.
}
}
