% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integerCounts.R
\name{integerCounts}
\alias{integerCounts}
\title{Integer counts}
\usage{
integerCounts(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
Matrix.
Typically \code{matrix} or \code{Matrix} class.
}
\description{
Integer counts
}
\note{
For a \code{SummarizedExperiment} object, \code{"counts"} must be explicitly
defined in \code{assayNames}.
}
\examples{
class(integerCounts)
}
