% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapGenesToIDs.R
\name{mapGenesToIDs}
\alias{mapGenesToIDs}
\title{Map genes (gene names) to gene identifiers}
\usage{
mapGenesToIDs(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{character}.
}
\description{
Map genes (gene names) to gene identifiers
}
\examples{
showMethods("mapGenesToIDs")
}
