% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match.R
\name{match}
\alias{match}
\alias{\%in\%}
\title{Value matching}
\usage{
match(x, table, nomatch = NA_integer_, incomparables = NULL, ...)

x %in% table
}
\arguments{
\item{x, table}{Vector-like objects (typically of the same class, but not necessarily).}

\item{nomatch, incomparables}{See [base::match][] for a description of these arguments.}

\item{...}{Additional arguments.}
}
\description{
Value matching
}
\note{
Updated 2020-12-21.
}
