% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchRowNameColumn.R
\name{matchRownameColumn}
\alias{matchRownameColumn}
\title{Match row name column}
\usage{
matchRownameColumn(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{character(1)}.
}
\description{
Match row name column
}
\examples{
showMethods("matchRownameColumn")
}
