% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/melt.R
\name{melt}
\alias{melt}
\title{Melt columns into key-value pairs}
\usage{
melt(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{DataFrame}.
}
\description{
Unpivot column data from wide format to long format.
}
\examples{
class(melt)
}
\seealso{
\itemize{
\item \code{tidyr::gather()}: Recommended instead for use with \code{tbl_df}.
\item \code{reshape2::melt()}: Inspiration for name, but now deprecated..
\item Python \code{pandas.melt}.
}
}
