% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meltCounts.R
\name{meltCounts}
\alias{meltCounts}
\title{Melt count matrix into long format}
\usage{
meltCounts(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{grouped_df}.
Grouped by \code{colname} (e.g. sample ID) and \code{rowname} (e.g. gene ID).
}
\description{
Melt count matrix into long format
}
\examples{
class(meltCounts)
}
\seealso{
\code{reshape2::melt()}.
}
