% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meltCounts.R
\name{meltCounts}
\alias{meltCounts}
\title{Melt count matrix into long format}
\usage{
meltCounts(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\description{
Melt count matrix into long format
}
