% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata2.R
\name{metadata2}
\alias{metadata2}
\alias{metadata2<-}
\title{Metadata}
\usage{
metadata2(x, which, ...)

metadata2(x, which, ...) <- value
}
\arguments{
\item{x}{Object.}

\item{which}{\code{character(1)}.
A non-empty character string specifying which attribute is to be accessed.
Note that positional \code{numeric} arguments are currently not allowed.}

\item{...}{Additional arguments.}

\item{value}{\code{ANY}.
Metadata values to assign into slot, defined by \code{which}.}
}
\value{
\itemize{
\item \code{metadata2()}: \code{list}. Returns \code{NULL} on \code{which} match failure, similar
to \code{\link[base:attr]{attr()}} and \code{\link[S4Vectors:Annotated-class]{metadata()}}.
\item \verb{metadata2<-()}: Modified object.
}
}
\description{
Dynamically handles metadata assignment and extraction in a similar fashion
for both S3 and S4 objects.
}
\section{S3 class}{


Internally slots into \code{\link[base:attributes]{attributes()}}.
Attribute is also accessible via \code{attr(object, which)}.
}

\section{S4 class}{


Requires that object extends \code{\link[S4Vectors:Annotated-class]{Annotated}} class,
which supports \code{\link[S4Vectors:Annotated-class]{metadata()}}.

Internally slots into \code{\link[S4Vectors:Annotated-class]{metadata()}}.
Attribute is also accessible via \code{metadata(object)[[which]]}.
}

\examples{
class(metadata2)
}
