% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCellTypesPerCluster.R
\name{plotCellTypesPerCluster}
\alias{plotCellTypesPerCluster}
\title{Plot cell types per cluster}
\usage{
plotCellTypesPerCluster(object, markers, ...)
}
\arguments{
\item{object}{Object.}

\item{markers}{Object containing gene markers.}

\item{...}{Additional arguments.}
}
\value{
Plot.
}
\description{
Plot cell types per cluster
}
\examples{
class(plotCellTypesPerCluster)
}
