% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCountsPerGene.R
\name{plotCountsPerGene}
\alias{plotCountsPerGene}
\title{Plot counts per gene}
\usage{
plotCountsPerGene(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\description{
Plot counts per gene
}
