% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDEGHeatmap.R
\name{plotDEGHeatmap}
\alias{plotDEGHeatmap}
\title{Differentially expressed gene heatmap}
\usage{
plotDEGHeatmap(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
Plot.
}
\description{
This function is an extension of \code{\link[=plotHeatmap]{plotHeatmap()}} that is optimized for
automatic handling differentially expressed genes, rather than requiring
manual input of a gene vector. All of the optional parameters for
\code{\link[=plotHeatmap]{plotHeatmap()}} are also supported by this function.
}
\details{
To adjust the annotation columns, modify the
\code{\link[SummarizedExperiment:SummarizedExperiment-class]{colData()}} of the \code{counts} argument, which
must contain a \code{SummarizedExperiment}.
}
\examples{
showMethods("plotDEGHeatmap")
}
