% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDEGPCA.R
\name{plotDEGPCA}
\alias{plotDEGPCA}
\title{Plot differentially expressed gene principal component analysis}
\usage{
plotDEGPCA(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
Plot.
}
\description{
This function is an extension of \code{\link[BiocGenerics:plotPCA]{plotPCA()}} that is
optimized for automatic handling of differentially expressed genes, rather
than requiring manual input of a gene vector or subset object.
}
\details{
To adjust the annotation columns, modify the
\code{\link[SummarizedExperiment:colData]{colData()}} of the \code{counts} argument, which
must contain/extend a \code{SummarizedExperiment}.
}
\examples{
class(plotDEGPCA)
}
