% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDEGStackedBar.R
\name{plotDEGStackedBar}
\alias{plotDEGStackedBar}
\title{Stacked bar plot of DEGs}
\usage{
plotDEGStackedBar(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{ggplot}.
}
\description{
Stacked bar plot of DEGs
}
\examples{
showMethods("plotDEGStackedBar")
}
