% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDEGUpset.R
\name{plotDEGUpset}
\alias{plotDEGUpset}
\title{UpSet plot of directional DEG intersections across contrasts}
\usage{
plotDEGUpset(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
Plot.
}
\description{
UpSet plot of directional DEG intersections across contrasts
}
\examples{
class(plotDEGUpset)
}
