% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGenderMarkers.R
\name{plotGenderMarkers}
\alias{plotGenderMarkers}
\title{Plot sexually dimorphic gender marker genes}
\usage{
plotGenderMarkers(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{ggplot}.
}
\description{
This is a convenience function that quickly plots known sexually dimorphic
genes, and overall gene experssion for the X and Y chromosomes.
}
\seealso{
\code{plotCounts()}.
}
