% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGeneEffectVsExpression.R
\name{plotGeneEffectVsExpression}
\alias{plotGeneEffectVsExpression}
\title{Plot gene effect vs. expression}
\usage{
plotGeneEffectVsExpression(effect, expression, ...)
}
\arguments{
\item{effect}{Object containing gene effect data.}

\item{expression}{Object containing gene expression data.}

\item{...}{Additional arguments.}
}
\value{
\code{ggplot}.
}
\description{
Plot gene effect vs. expression
}
\examples{
showMethods("plotGeneEffectVsExpression")
}
\seealso{
\itemize{
\item DepMapAnalysis package.
}
}
