% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGSEATable.R
\name{plotGSEATable}
\alias{plotGSEATable}
\title{Plot GSEA enrichment table}
\usage{
plotGSEATable(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\description{
Plot GSEA enrichment table
}
