% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotLfc.R
\name{plotLfc}
\alias{plotLfc}
\title{Plot log2 fold change distributions}
\usage{
plotLfc(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{ggplot}.
}
\description{
Plot log2 fold change distributions
}
\examples{
showMethods("plotLfc")
}
