% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMa.R
\name{plotMa}
\alias{plotMa}
\title{MA plot}
\usage{
plotMa(object, ...)
}
\value{
Plot.
}
\description{
MA plot
}
\note{
Updated 2023-09-27.
}
\examples{
showMethods("plotMa")
}
\seealso{
\itemize{
\item \code{BiocGenerics::plotMA}.
}
}
