% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioc-plotMA.R
\name{plotMA}
\alias{plotMA}
\title{MA plot}
\usage{
plotMA(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{ggplot}.
}
\description{
The plot visualizes the differences between measurements taken in two
samples, by transforming the data onto \strong{M} (log ratio) and \strong{A}
(mean average) scales.
}
\examples{
showMethods("plotMA")
}
