% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotNES.R
\name{plotNES}
\alias{plotNES}
\title{Plot normalized enrichment score}
\usage{
plotNES(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{ggplot}.
}
\description{
Plot normalized enrichment score
}
\examples{
class(plotNES)
}
