% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioconductor-02-generics.R
\name{plotPCA}
\alias{plotPCA}
\title{Principal component analysis plot}
\usage{
plotPCA(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{ggplot}.
}
\description{
Principal component analysis plot
}
\examples{
showMethods("plotPCA")
}
