% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPca.R
\name{plotPca}
\alias{plotPca}
\title{Principal component analysis plot}
\usage{
plotPca(object, ...)
}
\value{
Plot.
}
\description{
Principal component analysis plot
}
\note{
Updated 2023-09-27.
}
\examples{
showMethods("plotPca")
}
\seealso{
\itemize{
\item \code{BiocGenerics::plotPCA}.
}
}
