% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPCACovariates.R
\name{plotPCACovariates}
\alias{plotPCACovariates}
\title{Find correlation between principal components (PCs) and covariates}
\usage{
plotPCACovariates(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{ggplot}.
}
\description{
Find correlation between principal components (PCs) and covariates
}
\examples{
class(plotPCACovariates)
}
