% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPCElbow.R
\name{plotPCElbow}
\alias{plotPCElbow}
\title{Plot principal component elbow}
\usage{
plotPCElbow(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{ggplot}.
}
\description{
Calculate the principal component (PC) cutoff using a heuristic approach.
}
\examples{
class(plotPCElbow)
}
