% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotZerosVsDepth.R
\name{plotZerosVsDepth}
\alias{plotZerosVsDepth}
\title{Plot percentage of zeros vs. library depth}
\usage{
plotZerosVsDepth(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{ggplot}.
}
\description{
Plot percentage of zeros vs. library depth
}
\examples{
showMethods("plotZerosVsDepth")
}
