% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{combine}
\alias{organism}
\alias{organism<-}
\alias{plotMA}
\alias{plotPCA}
\alias{table}
\alias{unlist}
\alias{updateObject}
\alias{decode}
\alias{summary}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{BiocGenerics}{\code{\link[BiocGenerics]{combine}}, \code{\link[BiocGenerics:organism_species]{organism}}, \code{\link[BiocGenerics:organism_species]{organism<-}}, \code{\link[BiocGenerics]{plotMA}}, \code{\link[BiocGenerics]{plotPCA}}, \code{\link[BiocGenerics]{table}}, \code{\link[BiocGenerics]{unlist}}, \code{\link[BiocGenerics]{updateObject}}}

  \item{S4Vectors}{\code{\link[S4Vectors:Rle-class]{decode}}}
}}

