% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{sampleNames}
\alias{sampleNames<-}
\alias{\%in\%}
\alias{as.data.frame}
\alias{as.list}
\alias{colSums}
\alias{combine}
\alias{counts}
\alias{counts<-}
\alias{do.call}
\alias{estimateSizeFactors}
\alias{lapply}
\alias{match}
\alias{normalize}
\alias{order}
\alias{organism}
\alias{organism<-}
\alias{plotDispEsts}
\alias{plotMA}
\alias{plotPCA}
\alias{rowMeans}
\alias{rowSums}
\alias{setdiff}
\alias{sizeFactors}
\alias{sizeFactors<-}
\alias{t}
\alias{table}
\alias{unique}
\alias{unlist}
\alias{updateObject}
\alias{aggregate}
\alias{coerce}
\alias{complete.cases}
\alias{cor}
\alias{decode}
\alias{droplevels}
\alias{expand}
\alias{expand.grid}
\alias{head}
\alias{merge}
\alias{mcols}
\alias{mcols<-}
\alias{metadata}
\alias{metadata<-}
\alias{na.omit}
\alias{split}
\alias{summary}
\alias{tail}
\title{Objects exported from other packages}
\description{
These objects are imported from other packages.
}
\note{
Updated 2020-12-21.
}
\keyword{internal}
