% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{Annotated-class}
\alias{AsIs-class}
\alias{\%in\%}
\alias{colSums}
\alias{complete.cases}
\alias{cor}
\alias{do.call}
\alias{expand}
\alias{expand.grid}
\alias{head}
\alias{is.unsorted}
\alias{lapply}
\alias{match}
\alias{mcols}
\alias{mcols<-}
\alias{merge}
\alias{metadata}
\alias{metadata<-}
\alias{na.omit}
\alias{order}
\alias{rowMeans}
\alias{rowSums}
\alias{setdiff}
\alias{sort}
\alias{split}
\alias{summary}
\alias{t}
\alias{table}
\alias{tail}
\alias{unique}
\alias{unlist}
\alias{unsplit}
\alias{var}
\title{Re-exported S4 classes, generics, and methods}
\description{
Re-exported S4 classes, generics, and methods
}
\details{
Use \code{findMethod()} to locate the package where the methods are defined.
The \code{showMethod()} or \code{getMethod()} approaches don't clearly show package
for \code{derivedDefaultMethod}, which are defined in BiocGenerics.

Here's how to locate a method:\if{html}{\out{<div class="r">}}\preformatted{findMethod("do.call", signature(what="ANY", args="List"))
}\if{html}{\out{</div>}}
}
\keyword{internal}
