% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{sampleNames}
\alias{sampleNames<-}
\alias{\%in\%}
\alias{as.data.frame}
\alias{as.list}
\alias{colSums}
\alias{combine}
\alias{counts}
\alias{counts<-}
\alias{do.call}
\alias{estimateSizeFactors}
\alias{lapply}
\alias{match}
\alias{normalize}
\alias{order}
\alias{organism}
\alias{organism<-}
\alias{plotDispEsts}
\alias{plotMA}
\alias{plotPCA}
\alias{rowMeans}
\alias{rowSums}
\alias{setdiff}
\alias{sizeFactors}
\alias{sizeFactors<-}
\alias{t}
\alias{table}
\alias{unique}
\alias{unlist}
\alias{updateObject}
\alias{aggregate}
\alias{coerce}
\alias{complete.cases}
\alias{cor}
\alias{decode}
\alias{droplevels}
\alias{expand}
\alias{expand.grid}
\alias{head}
\alias{merge}
\alias{mcols}
\alias{mcols<-}
\alias{metadata}
\alias{metadata<-}
\alias{summary}
\alias{tail}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{Biobase}{\code{\link[Biobase:featureNames]{sampleNames}}, \code{\link[Biobase:featureNames]{sampleNames<-}}}

  \item{BiocGenerics}{\code{\link[BiocGenerics:match]{\%in\%}}, \code{\link[BiocGenerics]{as.data.frame}}, \code{\link[BiocGenerics]{as.list}}, \code{\link[BiocGenerics]{colSums}}, \code{\link[BiocGenerics]{combine}}, \code{\link[BiocGenerics:dge]{counts}}, \code{\link[BiocGenerics:dge]{counts<-}}, \code{\link[BiocGenerics]{do.call}}, \code{\link[BiocGenerics:dge]{estimateSizeFactors}}, \code{\link[BiocGenerics]{lapply}}, \code{\link[BiocGenerics]{match}}, \code{\link[BiocGenerics]{normalize}}, \code{\link[BiocGenerics]{order}}, \code{\link[BiocGenerics:organism_species]{organism}}, \code{\link[BiocGenerics:organism_species]{organism<-}}, \code{\link[BiocGenerics:dge]{plotDispEsts}}, \code{\link[BiocGenerics]{plotMA}}, \code{\link[BiocGenerics]{plotPCA}}, \code{\link[BiocGenerics:colSums]{rowMeans}}, \code{\link[BiocGenerics:colSums]{rowSums}}, \code{\link[BiocGenerics:sets]{setdiff}}, \code{\link[BiocGenerics:dge]{sizeFactors}}, \code{\link[BiocGenerics:dge]{sizeFactors<-}}, \code{\link[BiocGenerics]{t}}, \code{\link[BiocGenerics]{table}}, \code{\link[BiocGenerics]{unique}}, \code{\link[BiocGenerics]{unlist}}, \code{\link[BiocGenerics]{updateObject}}}
}}

