% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{as.data.frame}
\alias{combine}
\alias{counts}
\alias{counts<-}
\alias{estimateSizeFactors}
\alias{normalize}
\alias{organism}
\alias{organism<-}
\alias{plotDispEsts}
\alias{plotMA}
\alias{plotPCA}
\alias{rowMeans}
\alias{rowSums}
\alias{sizeFactors}
\alias{sizeFactors<-}
\alias{table}
\alias{unlist}
\alias{updateObject}
\alias{aggregate}
\alias{decode}
\alias{droplevels}
\alias{head}
\alias{merge}
\alias{summary}
\alias{tail}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{BiocGenerics}{\code{\link[BiocGenerics]{as.data.frame}}, \code{\link[BiocGenerics]{combine}}, \code{\link[BiocGenerics:dge]{counts}}, \code{\link[BiocGenerics:dge]{counts<-}}, \code{\link[BiocGenerics:dge]{estimateSizeFactors}}, \code{\link[BiocGenerics]{normalize}}, \code{\link[BiocGenerics:organism_species]{organism}}, \code{\link[BiocGenerics:organism_species]{organism<-}}, \code{\link[BiocGenerics:dge]{plotDispEsts}}, \code{\link[BiocGenerics]{plotMA}}, \code{\link[BiocGenerics]{plotPCA}}, \code{\link[BiocGenerics:colSums]{rowMeans}}, \code{\link[BiocGenerics:colSums]{rowSums}}, \code{\link[BiocGenerics:dge]{sizeFactors}}, \code{\link[BiocGenerics:dge]{sizeFactors<-}}, \code{\link[BiocGenerics]{table}}, \code{\link[BiocGenerics]{unlist}}, \code{\link[BiocGenerics]{updateObject}}}
}}

