% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioconductor-01-functions.R,
%   R/bioconductor-02-generics.R, R/bioconductor-03-classes.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{CharacterList}
\alias{DataFrame}
\alias{DataFrameList}
\alias{Factor}
\alias{FactorList}
\alias{IRanges}
\alias{IntegerList}
\alias{LLint}
\alias{List}
\alias{LogicalList}
\alias{NumericList}
\alias{Rle}
\alias{RleList}
\alias{SimpleList}
\alias{SplitDataFrameList}
\alias{\%in\%}
\alias{Filter}
\alias{Find}
\alias{Map}
\alias{Position}
\alias{Reduce}
\alias{anyDuplicated}
\alias{append}
\alias{as.factor}
\alias{as.matrix}
\alias{as.table}
\alias{as.vector}
\alias{basename}
\alias{cbind}
\alias{colSums}
\alias{colnames}
\alias{colnames<-}
\alias{complete.cases}
\alias{cor}
\alias{design}
\alias{dims}
\alias{dirname}
\alias{do.call}
\alias{duplicated}
\alias{end}
\alias{end<-}
\alias{estimateDispersions}
\alias{eval}
\alias{expand}
\alias{expand.grid}
\alias{get}
\alias{getListElement}
\alias{grep}
\alias{grepl}
\alias{gsub}
\alias{intersect}
\alias{is.unsorted}
\alias{lapply}
\alias{mapply}
\alias{match}
\alias{mcols}
\alias{mcols<-}
\alias{median}
\alias{mean}
\alias{merge}
\alias{metadata}
\alias{metadata<-}
\alias{mget}
\alias{na.omit}
\alias{ncol}
\alias{nrow}
\alias{order}
\alias{paste}
\alias{pmax}
\alias{pmax.int}
\alias{pmin}
\alias{pmin.int}
\alias{pos}
\alias{quantile}
\alias{ranges}
\alias{rank}
\alias{rbind}
\alias{relist}
\alias{rep.int}
\alias{rowMeans}
\alias{rowSums}
\alias{rownames}
\alias{rownames<-}
\alias{sapply}
\alias{sd}
\alias{setdiff}
\alias{sort}
\alias{split}
\alias{start}
\alias{start<-}
\alias{sub}
\alias{subset}
\alias{t}
\alias{table}
\alias{tapply}
\alias{trim}
\alias{union}
\alias{unique}
\alias{unlist}
\alias{unsplit}
\alias{var}
\alias{weights}
\alias{which}
\alias{which.max}
\alias{which.min}
\alias{width}
\alias{width<-}
\alias{Annotated-class}
\alias{AsIs-class}
\alias{AtomicList-class}
\alias{CharacterList-class}
\alias{DataFrame-class}
\alias{DataFrameList-class}
\alias{Factor-class}
\alias{FactorList-class}
\alias{IRanges-class}
\alias{IntegerList-class}
\alias{LLint-class}
\alias{List-class}
\alias{LogicalList-class}
\alias{NumericList-class}
\alias{RectangularData-class}
\alias{Rle-class}
\alias{RleList-class}
\alias{SimpleList-class}
\alias{SplitDataFrameList-class}
\alias{Vector-class}
\alias{atomic-class}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{IRanges}{\code{\link[IRanges:AtomicList-class]{CharacterList}}, \code{\link[IRanges:DataFrameList-class]{DataFrameList}}, \code{\link[IRanges:AtomicList-class]{FactorList}}, \code{\link[IRanges:IRanges-constructor]{IRanges}}, \code{\link[IRanges:AtomicList-class]{IntegerList}}, \code{\link[IRanges:AtomicList-class]{LogicalList}}, \code{\link[IRanges:AtomicList-class]{NumericList}}, \code{\link[IRanges:AtomicList-class]{RleList}}, \code{\link[IRanges:DataFrameList-class]{SplitDataFrameList}}}

  \item{S4Vectors}{\code{\link[S4Vectors:DataFrame-class]{DataFrame}}, \code{\link[S4Vectors:Factor-class]{Factor}}, \code{\link[S4Vectors:LLint-class]{LLint}}, \code{\link[S4Vectors:List-class]{List}}, \code{\link[S4Vectors:Rle-class]{Rle}}, \code{\link[S4Vectors:SimpleList-class]{SimpleList}}}
}}

