% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioc-reexports-classes.R, R/bioc-reexports.R
\docType{import}
\name{Annotated-class}
\alias{Annotated-class}
\alias{AsIs-class}
\alias{DataFrame-class}
\alias{Factor-class}
\alias{LLint-class}
\alias{List-class}
\alias{RectangularData-class}
\alias{Rle-class}
\alias{SimpleList-class}
\alias{Vector-class}
\alias{atomic-class}
\alias{reexports}
\alias{DataFrame}
\alias{Factor}
\alias{LLint}
\alias{List}
\alias{Rle}
\alias{SimpleList}
\alias{\%in\%}
\alias{cbind}
\alias{colSums}
\alias{complete.cases}
\alias{cor}
\alias{do.call}
\alias{expand}
\alias{expand.grid}
\alias{is.unsorted}
\alias{lapply}
\alias{match}
\alias{mcols}
\alias{mcols<-}
\alias{merge}
\alias{metadata}
\alias{metadata<-}
\alias{na.omit}
\alias{order}
\alias{rbind}
\alias{rowMeans}
\alias{rowSums}
\alias{setdiff}
\alias{sort}
\alias{split}
\alias{t}
\alias{table}
\alias{unique}
\alias{unlist}
\alias{unsplit}
\alias{var}
\title{Re-exported S4 classes, generics, and methods}
\details{
Use \code{findMethod()} to locate the package where the methods are defined.
The \code{showMethod()} or \code{getMethod()} approaches don't clearly show package
for \code{derivedDefaultMethod}, which are defined in BiocGenerics.

Here's how to locate a method:\if{html}{\out{<div class="r">}}\preformatted{findMethod("do.call", signature(what="ANY", args="List"))
}\if{html}{\out{</div>}}
}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{S4Vectors}{\code{\link[S4Vectors:DataFrame-class]{DataFrame}}, \code{\link[S4Vectors:Factor-class]{Factor}}, \code{\link[S4Vectors:LLint-class]{LLint}}, \code{\link[S4Vectors:List-class]{List}}, \code{\link[S4Vectors:Rle-class]{Rle}}, \code{\link[S4Vectors:SimpleList-class]{SimpleList}}}
}}

