% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relativeLogExpression.R
\name{relativeLogExpression}
\alias{relativeLogExpression}
\title{Relative log expression}
\usage{
relativeLogExpression(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{matrix}.
}
\description{
Relative log expression
}
\examples{
showMethods("relativeLogExpression")
}
\references{
Anders and Huber (2010).
}
\seealso{
\code{edgeR::calcNormFactors()}.
}
