% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeNa.R
\name{removeNa}
\alias{removeNa}
\title{Remove rows and columns containing only NA values}
\usage{
removeNa(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
Sanitized object.
}
\description{
Remove rows and columns containing only NA values
}
\examples{
showMethods("removeNa")
}
