% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resultsNames.R
\name{resultsNames}
\alias{resultsNames}
\title{Results names}
\usage{
resultsNames(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{character}.
}
\description{
Results names
}
\seealso{
\itemize{
\item \code{DESeq2::resultsNames()}.
}
}
