% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanitizeNA.R
\name{sanitizeNA}
\alias{sanitizeNA}
\title{Sanitize \code{NA} values}
\usage{
sanitizeNA(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\description{
Sanitize \code{NA} values
}
