% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanitizePercent.R
\name{sanitizePercent}
\alias{sanitizePercent}
\title{Sanitize percentage}
\usage{
sanitizePercent(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
Modified object, with all percentage strings containing \code{"\%"} (e.g.
\code{"50\%"}) converted to numeric proportions (e.g. \code{0.5}).
}
\description{
Take a \code{character} vector of percentages (e.g. \code{"50\%"}) and convert it to a
\code{numeric} vector (e.g. \code{0.5}). This function is designed primarily to
sanitize data imported from Microsoft Excel.
}
