% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{select}
\alias{select}
\alias{selectIf}
\title{Select multiple columns}
\usage{
selectIf(object, predicate, ...)
}
\arguments{
\item{object}{Object.}

\item{predicate}{\code{function} or \code{logical}.
A predicate function to be applied to the columns, or a logical vector
matching the number of columns.}

\item{...}{Additional arguments.}
}
\value{
Modified object.
}
\description{
Select multiple columns
}
\note{
Updated 2019-08-26.
}
\seealso{
These functions are inspired by dplyr. However, they are designed to only
work on \code{DataFrame} class, and use base R code internally.\preformatted{`help(topic = "select_all", package = "dplyr")`
}
}
