% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectSamples.R
\name{selectSamples}
\alias{selectSamples}
\title{Select samples}
\usage{
selectSamples(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Selection arguments:
\itemize{
\item Key value pairs that must map to the column names of \code{\link[=sampleData]{sampleData()}}.
\item \code{atomic} values are supported.
\item Avoid using \code{logical} or \code{numeric} indices
(e.g. \code{\link[base:which]{which()}} return), thereby making your code more
human readable, and easier to interpret.
}}
}
\value{
Modified object.
Subset (extraction) containing the desired samples.
}
\description{
Utility function that enables quick and easy sample selection.
}
\details{
Internally, pattern matching against sample and file names is applied using
logical grep matching.
}
\section{Intended usage}{


This function is intended to be used for sample extraction from Bioconductor
container classes where the column names don't correspond to samples (e.g.
\code{SingleCellExperiment}).
}

\section{Bracket-based subsetting}{


Bracket-based subsetting with \code{[} also works on \code{SingleCellExperiment}
objects but it's not intuitive. In this case, provide cellular barcode
identifiers for columns and gene identifiers for rows.
}

\seealso{
\itemize{
\item \code{sampleData()}.
\item \code{S4Vectors::split()}.
}
}
