% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sizeFactors.R
\name{sizeFactors}
\alias{sizeFactors}
\alias{sizeFactors<-}
\title{Size factors}
\usage{
sizeFactors(object, ...)
sizeFactors(object, ...) <- value
}
\arguments{
\item{object}{Object.}

\item{value}{Value to assign.}

\item{...}{Additional arguments.}
}
\value{
Varies, depending on the method.
}
\description{
Size factors
}
\examples{
showMethods("sizeFactors")
}
