% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stripTranscriptVersions.R
\name{stripTranscriptVersions}
\alias{stripTranscriptVersions}
\title{Strip transcript versions}
\usage{
stripTranscriptVersions(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\description{
Strip transcript versions
}
