% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetPerSample.R
\name{subsetPerSample}
\alias{subsetPerSample}
\title{Subset per sample}
\usage{
subsetPerSample(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\description{
Subset per sample
}
