% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topCellsPerSample.R
\name{topCellsPerSample}
\alias{topCellsPerSample}
\title{Top cells per sample}
\usage{
topCellsPerSample(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{list}.
Top barcodes as \code{character}, split by \code{sampleID}.
}
\description{
Obtain the top cellular barcodes, based on counts.
}
\examples{
class(topCellsPerSample)
}
