% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcriptIDs.R
\name{transcriptIDs}
\alias{transcriptIDs}
\title{Transcript identifiers}
\usage{
transcriptIDs(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{character}.
}
\description{
Convenience function for quickly returning the stable transcript identifiers.
}
\examples{
showMethods("transcriptIDs")
}
