% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniteInterestingGroups.R
\name{uniteInterestingGroups}
\alias{uniteInterestingGroups}
\title{Unite interesting groups into a single column}
\usage{
uniteInterestingGroups(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\description{
Unite interesting groups into a single column
}
