% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unlistToDataFrame.R
\name{unlistToDataFrame}
\alias{unlistToDataFrame}
\title{Unlist to a data frame}
\usage{
unlistToDataFrame(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{DataFrame}.
}
\description{
Unlist to a data frame
}
\examples{
showMethods("unlistToDataFrame")
}
